% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{add_effect_sizes}
\alias{add_effect_sizes}
\title{Effect Sizes for Analysis of Variance}
\usage{
add_effect_sizes(x, es = "ges", observed = NULL, intercept = FALSE)
}
\arguments{
\item{x}{An object of class \code{apa_variance_table}.}

\item{es}{Character. A vector naming all to-be-computed effect-size measures.
Currently, partial eta-squared (\code{"pes"}), generalized eta-squared
(\code{"ges"}), and eta-squared (\code{"es"}) are supported.}

\item{observed}{Character. A vector naming all factors that are observed
(i.e., \emph{not} manipulated).}

\item{intercept}{Logical. Should the sum of squares of the intercept (i.e., the
deviation of the grand mean from 0) be included in the calculation of eta-squared?}
}
\description{
Calculates effect-size measures for Analysis of Variance output objects.
\emph{This function is not exported and will soon be deprecated.}
}
\keyword{internal}
