% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_equals}
\alias{add_equals}
\title{Add Equals Where Necessary}
\usage{
add_equals(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
This is an internal function that prepends every element of a character
vector with an 'equals' sign if the respective element does not contain one of
\code{c("=", "<", ">")}.
}
\keyword{internal}
