% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_table.R
\name{add_grouping_colnames}
\alias{add_grouping_colnames}
\title{Add table headings to group columns}
\usage{
add_grouping_colnames(table_lines, grouping_colnames, n_cols)
}
\arguments{
\item{table_lines}{Character. Vector of characters containing one line of a LaTeX table each.}

\item{grouping_colnames}{List. A named list containing the indices of the first and last columns to group, where the names are the headings.}

\item{n_cols}{Numeric. Number of columns of the table.}
}
\description{
Takes a named list of containing column numbers to group with a heading
\emph{This function is not exported.}
}
\examples{
NULL
}
\seealso{
\code{\link{apa_table}}
}

