% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{annotate}
\alias{annotate}
\alias{annotate,data.frame-method}
\title{Coerce vectors to annotated vectors}
\usage{
annotate(object)

\S4method{annotate}{data.frame}(object)
}
\arguments{
\item{An}{annotated vector or a data.frame.}
}
\description{
Internal workhorse function wthat will soon be dropped.
}
\keyword{internal}
