% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{class}
\name{annotated_numeric-class}
\alias{annotated_numeric-class}
\title{S4 class union 'annotated_vector'}
\description{
These are the details about this class..
}
\section{Slots}{

\describe{
\item{\code{label}}{Length-one character. Should be a copy of slot "label" in slot "annotation".
This slot is only included for nice printing in RStudio.}

\item{\code{annotation}}{An object of class \code{\link{vector_annotation-class}}, containing all meta-data of the vector.}
}}

\keyword{internal}
