% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{class}
\name{annotated_numeric-class}
\alias{annotated_numeric-class}
\alias{annotated_integer-class}
\alias{annotated_character-class}
\alias{annotated_logical-class}
\alias{annotated_factor-class}
\alias{annotated_named_numeric-class}
\alias{annotated_named_integer-class}
\alias{annotated_named_character-class}
\alias{annotated_named_logical-class}
\alias{annotated_named_factor-class}
\alias{annotated_named_vector-class}
\alias{annotated_vector-class}
\title{S4 Class Union for Annotated Vectors}
\description{
This is a collection of classes that extend basic data types (currently
vector_annotation-object.
}
\details{
We currently support six different classes of annotated vectors:
  \code{annotated_numeric}, \code{annotated_integer},
  \code{annotated_character}, \code{annotated_logical},
  \code{annotated_factor}.

  These can be extended to named vectors (via \code{\link{names}(x)}).
  The corresponding classes are
  \code{annotated_named_numeric}, \code{annotated_named_integer},
  \code{annotated_named_character}, \code{annotated_named_logical},
  \code{annotated_named_factor}.

  \emph{Ordered factors, raw and complex vectors are not supported, yet.}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Depending on the data type that the class extends, the original
values.}

\item{\code{names}}{A character vector of value names, if supplied.}

\item{\code{label}}{Length-one character. This one must always be a copy of slot
"label" in slot "annotation".
\emph{Don't you ever manipulate this one by hand.}
It is only included for nice printing in RStudio.}

\item{\code{annotation}}{A \code{\link{vector_annotation-class}} object,
containing all meta-data of the vector.}
}}

\keyword{internal}
