% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_article.R
\name{apa_article}
\alias{apa_article}
\title{APA article (6th edition, PDF)}
\usage{
apa_article(class = "man", fig.caption = TRUE, ...)
}
\arguments{
\item{class}{Character. Specifies if the document will be a manuscript ("man"), a document ("doc"),
or a fully typsetted journal article ("jou"). Currently only manuscripts are supported for
MS Word documents.}

\item{fig.caption}{Logical. Indicates if figures are rendered with captions.}

\item{...}{Further arguments to pass to \code{rmarkdown::pdf_document} or \code{rmarkdown::word_document}.}
}
\description{
Template for creating an article according to APA guidelines (6th edition) in PDF format.
}
\examples{
NULL
}

