% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_before_body.R
\name{apa_before_body}
\alias{apa_before_body}
\title{Prepare APA document}
\usage{
apa_before_body(x = "default")
}
\arguments{
\item{x}{Character. Document language to be set. Ignored and replaced by yaml-header parameter \code{lang}
when document is rendered.}
}
\description{
Prepares objekts to be used in the rendering process and creates title page and abstract in MS Word
documents.
}
\details{
The function creates and locks the object \code{apa_lang} in the global environment that is
   needed by other \pkg{papaja}-functions. \code{apa_lang} is a \code{list} containing localizations for
   document elements such as abstract title. The selected language is defined by the \code{lang}-parameter
   in the documents yaml-header. Currently, English (default) and German ("german") are supported.

   If the output document is MS Word (\code{output: \link{apa6_word}}) the function additionally creates
   a title page and adds the abstract before the document body.
}
\examples{
apa_before_body()
}
\seealso{
\code{\link{apa6_word}}
}

