% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_factorial_plot.R
\name{apa_factorial_plot_single}
\alias{apa_factorial_plot_single}
\title{Plots for factorial designs that conform to APA guidelines, two-factors internal function}
\usage{
apa_factorial_plot_single(
  aggregated,
  y.values,
  id,
  dv,
  factors,
  intercept = NULL,
  ...
)
}
\arguments{
\item{aggregated}{A \code{data.frame}, the \emph{aggregated} data.}

\item{y.values}{A \code{data.frame} containing the measures of central tendency and of dispersion per cell of the design.}

\item{id}{Character. Variable name that identifies subjects.}

\item{dv}{Character. The name of the dependent variable.}

\item{factors}{Character. A vector of up to four variable names that is used to stratify the data.}

\item{intercept}{Numeric. See details in \code{\link{apa_factorial_plot}}}
}
\description{
Internal function that is called (possibly multiple times) by \code{\link{apa_factorial_plot}}.
}
\keyword{internal}
