% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_plot.R
\name{apa_plot}
\alias{apa_plot}
\alias{apa_plot.formula}
\alias{apa_plot.data.frame}
\title{Plots that Conform to APA Guidelines}
\usage{
apa_plot(x, ...)

\method{apa_plot}{formula}(formula, data, ...)

\method{apa_plot}{data.frame}(
  data,
  dv,
  id = NULL,
  factors = NULL,
  use = "all.obs",
  tendency = mean,
  dispersion = conf_int,
  level = 0.95,
  fun_aggregate = mean,
  jit = 0.3,
  mfrow = NULL,
  reference = 0,
  ...
)
}
\arguments{
\item{x}{An object of class \code{formula} or \code{data.frame}}

\item{formula}{A two-sided formula specifying the dependent variable on the left
of a \code{~} operator, and factors on the right. A vertical bar (\code{|}) can be used to
specify an additional subject identifier, which must be provided if an additional
aggregation step is necessary. See the examples section for details.}

\item{data}{A data frame (or list) from which the variables in \code{formula} should be taken.}
}
\description{
Internal workhorse function for all \verb{apa_***plot()} functions.
}
\examples{
apa_plot(yield ~ (1 | block), data = npk) |>
 x_axis() |>
 y_axis() |>
 annotation() |>
 swarms() |>
 error_bars() |>
 points()
apa_plot(yield ~ (N | block), data = npk) |>
 x_axis() |>
 y_axis() |>
 annotation() |>
 swarms() |>
 error_bars() |>
 points()
apa_plot(yield ~ (N * P * K| block), data = npk) |>
 x_axis() |>
 y_axis() |>
 annotation() |>
 swarms() |>
 error_bars() |>
 points()
 \dontrun{
 # Both left-hand side and right-hand side must be provided:
 apa_plot(~ N * P)
 # One and only one identifier variable has to be provided:
 apa_plot(yield ~ 1)
 apa_plot(yield ~ (1|N*P), data = npk)
 }

}
