% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_prepare_doc.R
\name{apa_prepare_doc}
\alias{apa_prepare_doc}
\title{Prepare APA document}
\usage{
apa_prepare_doc(lang = "english")
}
\arguments{
\item{x}{Character. Document language to set. Ignored and replaced by yaml-header parameter \code{lang}
when document is rendered.}
}
\description{
Prepares objekts to be used in the rendering process and creates title page and abstract in MS Word
documents.
}
\details{
The function creates and locks a non-exported object \code{apa_lang} that is used by other
   \pkg{papaja}-functions. \code{apa_lang} is a \code{list} containing localizations for document
   elements such as abstract title. The selected language is defined by the \code{lang}-parameter
   in the documents yaml-header. Currently, English (default) and German ("german") are supported.

   If the output document is MS Word (\code{output: \link{apa6_word}}) the function additionally creates
   a title page and adds the abstract.
}
\examples{
apa_prepare_doc()
}
\seealso{
\code{\link{apa6_word}}
}

