% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_print_anova.R
\name{apa_print.anova}
\alias{apa_print.Anova.mlm}
\alias{apa_print.anova}
\title{Format statistics (APA 6th edition) from ANOVAs}
\usage{
\method{apa_print}{anova}(x, ...)

\method{apa_print}{Anova.mlm}(x, correction = "GG", ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{correction}{\code{character} In the case of repeated-measures ANOVA, the type of sphericity correction to be used. Either "GG" for Greenhouse-Geisser or "HF" for Huyn-Feldt methods. "none" is also possible.}

\item{es}{\code{character} The effect-size measure to be calculated. Either "ges" for generalized eta-squared or "pes" for partial eta-squared.}

\item{observed}{\code{character} The names of the factors that are observed, (i.e., not manipulated). Is necessary for calculation of generalized eta-squared.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}
}
\value{
a named list
}
\description{
These methods take input objects of the \code{anova} type and return them in a format that is suitable for in-line printing.
}
\details{
details
}
\examples{
NULL
}
\seealso{
Other apa_print: \code{\link{apa_print.aov}},
  \code{\link{apa_print.aovlist}},
  \code{\link{apa_print.summary.aov}};
  \code{\link{apa_print.htest}};
  \code{\link{apa_print.lm}},
  \code{\link{apa_print.summary.lm}};
  \code{\link{apa_print}}
}

