% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anova.R
\name{apa_print.aovlist}
\alias{apa_print.aovlist}
\alias{apa_print.summary.aov}
\title{Format statistics (APA 6th edition)}
\usage{
\method{apa_print}{aovlist}(x, ...)

\method{apa_print}{summary.aov}(x, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{es}{\code{character} The effect-size measure to be calculated. Either "ges" for generalized eta-squared or "pes" for partial eta-squared.}

\item{observed}{\code{character} The names of the factors that are observed, (i.e., not manipulated). Is necessary for calculation of generalized eta-squared.}

\item{op}{}

\item{cp}{}

\item{correction}{\code{character} In the case of repeated-measures ANOVA, the type of sphericity correction to be used. Either "GG" for Greenhouse-Geisser or "HF" for Huyn-Feldt methods. "none" is also possible.}
}
\value{
a named list
}
\description{
\code{apa_print.aov}
\code{apa_print.aovlist}
\code{apa_print.summary.aov}
\code{apa_print.anova}
\code{apa_print.Anova.mlm}
}
\details{
details
}
\examples{
NULL
}
\seealso{
Other apa_print: \code{\link{apa_print.htest}};
  \code{\link{apa_print.lm}},
  \code{\link{apa_print.summary.lm}};
  \code{\link{apa_print}}
}

