% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_stat.R
\name{apa_stat}
\alias{apa_stat}
\title{Format statistics (APA 6th edition)}
\usage{
apa_stat(x, stat_name = NULL, n = NULL, standardized = FALSE, ci = NULL,
  in_paren = FALSE)
}
\arguments{
\item{x}{Output object. See details.}

\item{stat_name}{Character. Name of the reported statistic; if \code{NULL} the name specified in the
supplied object is used.}

\item{n}{Numeric. Size of the sample. Needed when reporting $\chi^2$, otherwise this parameter is
ignored.}

\item{standardized}{Logical. If \code{x} is an object of class \code{summary.lm} indicates if
coefficients are standardized or unstandardized.}

\item{ci}{Numeric. Matrix containing lower and upper boundaries of confidence intervals for estimated
coefficients in columns. Ignored if \code{x} is not of class \code{summary.lm}.}

\item{in_paren}{Logical. Indicates if the formated string be reported inside parentheses. See details.}
}
\description{
Formats output objects from various statistical methods to create formated chraracter strings to report
the results in accordance with APA manuscript guidelines.
}
\details{
Currently the following output objects are supported:

   \itemize{
     \item \code{htest}
     \item \code{summary.lm}
     \item \code{anova}
   }

   If \code{in_paren} is \code{TRUE} parentheses in the formated string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
t_stat <- t.test(iris$Sepal.Length, iris$Sepal.Width)
apa_stat(t_stat)

cor_stat <- cor.test(iris$Sepal.Length, iris$Sepal.Width)
apa_stat(cor_stat)

iris_lm <- lm(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
apa_stat(summary(iris_lm), ci = confint(iris_lm))

iris_lm2 <- update(iris_lm, formula = . ~ + Sepal.Length:Sepal.Width)
apa_stat(anova(iris_lm2, iris_lm))
}

