% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_table.R
\name{apa_table}
\alias{apa_table}
\title{Prepare table for printing}
\usage{
apa_table(...)
}
\arguments{
\item{...}{Further arguments to pass to \code{kntir::kable}.}

\item{x}{Object to print, can be \code{matrix}, \code{data.frame}, or \code{list}. See details.}

\item{caption}{Character. Caption to be printed above the table.}

\item{note}{Character. Note to be printed below the table.}

\item{placement}{Character. Indicates wheter table should be placed at the exact location ("h"), at the
top ("t"), bottom ("b"), or on a new page ("p"). Arguments can be combined ("htb"); ignored in MS
Word documents.}

\item{landscape}{Logical. If \code{TRUE} the table is printed in landscape format. Ignored in MS Word
documents.}

\item{added_colnames}{Character. Vector of names for first unnamed columns. See details.}
}
\description{
Formats \code{matrices} and \code{data.frames} to report them as tables according to APA guidelines
(6th edition).
}
\details{
If \code{x} is a \code{list}, all list elements are merged by columns into a single table with
   the first column giving the names of elements.

   If the first column(s) of the table are unnamed, the names for these columns can be supplied using the
   \code{added_colnames} parameter. This is the case when an object has rownames (unless
   \code{row.names = FALSE} is passed to \code{knitr::kable}) and when elements of a \code{list} are
   merged.

   The syntax of the ouput is determined automatically by the rendered document type. If no rendering is
   in progress the output default is LaTeX.
}
\examples{
...
}
\seealso{
\code{\link{apply}}
}

