% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_plots.R
\name{arrows.matrix}
\alias{arrows.matrix}
\title{Matrix method for arrows}
\usage{
arrows.matrix(x0, x1, y0, y1, ...)
}
\description{
Internal function for convenient plotting of multiple arrows.
}
\keyword{internal}
