% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{as.character.papaja_labelled}
\alias{as.character.papaja_labelled}
\title{Conversion of Labelled Vectors}
\usage{
\method{as.character}{papaja_labelled}(x, keep_label = TRUE, ...)
}
\arguments{
\item{x}{Object to be coerced}

\item{keep_label}{Logical indicating whether the variable labels should be kept.}

\item{...}{Further arguments passed to or from methods}
}
\description{
Functions to convert labelled vectors to other representations.
}
