% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{as.character,annotated_factor-method}
\alias{as.character,annotated_factor-method}
\title{Coerce annotated factors to character vectors}
\usage{
\S4method{as.character}{annotated_factor}(x)
}
\arguments{
\item{x}{An object of class annotated_factor.}
}
\description{
Methods to coerce an annotated factor to an atomic character vector.
Necessary for compatibility reasons (View() in RStudio).
}
\keyword{internal}
