% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{as.factor,annotated_vector-method}
\alias{as.factor,annotated_vector-method}
\title{Coerce Annotated Vectors to Factors}
\usage{
\S4method{as.factor}{annotated_vector}(x)
}
\arguments{
\item{x}{An object of virtual class annotated_vector.}
}
\description{
Methods to coerce an annotated vector to an (atomic) factor.
Necessary for compatibility reasons.
}
\keyword{internal}
