% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canonize}
\alias{canonize}
\title{Transform to a Canonical Table}
\usage{
canonize(x, stat_label = NULL, est_label = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{stat_label}{Character. Label for column \code{statistic}.}

\item{est_label}{Character. Label for column \code{estimate}.}
}
\description{
Internal function that transforms a data frame by renaming and labelling columns.
}
\keyword{internal}
