% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_barplot.R
\name{conf.int}
\alias{conf.int}
\title{Between-subjects confidence intervals}
\usage{
conf.int(x, level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{\code{numeric}. A vector of observations from your dependent variable.}

\item{level}{\code{numeric}. The area of the cdf that will be covered. Defaults to .95 for 95\% confidence intervals.}

\item{na.rm}{\code{logical}. Shall missing values (i.e., \code{NA}s) be removed?}
}
\description{
returns confidence intervals
}

