% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_stat_name}
\alias{convert_stat_name}
\title{Convert name of statistic}
\usage{
convert_stat_name(x)
}
\arguments{
\item{x}{Chracter.}
}
\description{
This function converts a character generated by R-functions that describes a statistic and converts it into the
corresponding character required by APA guidelines (6th edition). \emph{This function is not exported.}
}
\examples{
\dontrun{
convert_stat_name("rho")
convert_stat_name("mean of the differences")
convert_stat_name("t")
}
}

