% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corresponding_author_line}
\alias{corresponding_author_line}
\title{Corresponding author line}
\usage{
corresponding_author_line(x)
}
\arguments{
\item{x}{List. Meta data of the document as a result from \code{\link[yaml]{yaml.load}}.}
}
\description{
Internal function. Construct corresponding-author line.
}
\keyword{internal}
