% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_container}
\alias{create_container}
\title{Create APA Results List from a Beautified Table}
\usage{
create_container(x, in_paren, add_par = NULL, sanitized_terms = NULL)
}
\arguments{
\item{in_paren}{Logical. Indicates if parentheses in the output strings (e.g., thos around degrees of freedom)
should be replaced with brackets. Useful if the complete output strings are supposed to be reported within parentheses.}
}
\description{
Internal function that takes a beautified table and constructs an object of class
\code{c("apa_results", "list")}, the canonical output object from \code{\link{apa_print}}.
}
\keyword{internal}
