% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_lineplot.R
\name{defaults}
\alias{defaults}
\title{Set defaults}
\usage{
defaults(ellipsis, set = NULL, set.if.null = NULL)
}
\arguments{
\item{ellipsis}{A \code{list}, usually a list that comes from an ellipsis}

\item{set}{A named  \code{list} of parameters that are intended to be set.}

\item{set.if.null}{A named \code{list} of parameters that are intended to be set if and only if the parameter is not already in \code{ellipsis}.}
}
\description{
A helper function that is intended for internal use. A list \code{ellipsis} may be manipulated by overwriting (via \code{set}) or adding (via \code{set.if.null}) list elements.
}

