% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{droplevels,data.frame-method}
\alias{droplevels,data.frame-method}
\title{#' Drop Unused Levels from Vectors
#'
#' This is an additional method that allows for more convenient programming:
#' If a vector (annotated or not) is not a factor, do not change anything.
#' @param x An atomic vector or annotated vector.
#' @return x, unchanged.
#' @keywords internal}
\usage{
\S4method{droplevels}{data.frame}(x, except = NULL, exclude)
}
\arguments{
\item{x}{An atomic vector or annotated vector.}
}
\value{
x, unchanged.
}
\description{
setMethod(
  f = "droplevels"
  , signature = "vector"
  , definition = function(x) {
    x
  }
)
}
\details{
Drop Unused Levels from Data Frames

This is a copy from the S3-method from the \pkg{base} package.
}
