% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{droplevels,vector-method}
\alias{droplevels,vector-method}
\title{Drop Unused Levels}
\usage{
\S4method{droplevels}{vector}(x)
}
\arguments{
\item{x}{An atomic vector or annotated vector.}
}
\value{
x, unchanged.
}
\description{
This is an additional method that allows for more convenient programming:
If a vector is not a factor (annotated or not), do not change anything.
}
\keyword{internal}
