% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{droplevels.labelled}
\alias{droplevels.labelled}
\title{Drop unused levels of factor}
\usage{
\method{droplevels}{labelled}(x, exclude = if (anyNA(levels(x))) NULL else NA,
  ...)
}
\description{
This one is intended to become the S3 method for labelled factors. Unfortunately, this only works reliably for R >= 3.4.0.
Therefore, we postponed exporting this function.
}
\keyword{internal}
