% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape Symbols for LaTeX Output}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{Character.}

\item{newlines}{Logical. Determines if \code{\\n} are escaped.}

\item{spaces}{Logical. Determines if multiple spaces are escaped.}
}
\description{
This function is a copy of the non-exported function \code{escape_latex()} from the \pkg{knitr} package.
\emph{This function is not exported.}
}
\keyword{internal}
