% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{factor.labelled}
\alias{factor.labelled}
\title{Provide a method for factor}
\usage{
\method{factor}{labelled}(x, ...)
}
\description{
It would be nice to export this as a method
This one is necessary for the S3 methods of droplevels, relevel, reorder, as.data.frame
}
\keyword{internal}
