% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{fast_aggregate}
\alias{fast_aggregate}
\title{Aggregate data much faster using dplyr}
\usage{
fast_aggregate(data, factors, dv, fun, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data.}

\item{factors}{Character. Data is intended to be aggregated by these factors.}

\item{dv}{Character. The dependent variable that is aggregated. All variables in \code{data} that contain this string will be aggregated separately.}

\item{fun}{The function used for aggregation.}
}
\description{
This is a convenience wrapper for aggregating your data using dplyr functions that tend to be much faster than the usual \code{aggregate} command.
It is also easy to call from within a function.
}
\details{
If \code{x} is a vector, \code{digits}, \code{gt1}, and \code{zero} can be vectors
   according to which each element of the vector is formated. Parameters are recycled if length of \code{x}
   exceeds length of the parameter vectors. If \code{x} is a matrix, the vectors specify the formating
   of either rows or columns according to the value of \code{margin}.
}
\examples{
NULL
}

