% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{fetch_web_refs}
\alias{fetch_web_refs}
\title{Fetch a .bib-reference file from the web (defunct)}
\usage{
fetch_web_refs(x, bib_name)
}
\arguments{
\item{x}{Character. URL of the \code{.bib}-file to fetch.}

\item{bib_name}{Character. The path and name of the file to be created.}
}
\description{
Downloads and saves a \code{.bib}-reference file form the web, so it can be used to cite references
in a Markdown-document using pandoc or LaTeX. \emph{This function has been defunct. Please use
\code{\link[downloader]{download}} instead.}
}
\details{
If the function is called in an RMarkdown-document the file name specified as \code{bib_name}
   can be used in the YAML header as \code{bibliography}
}
\examples{
NULL
}
\seealso{
\code{\link{cite_r}}, \code{\link{r_refs}}, \code{\link[knitr]{write_bib}}
}
