% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{format_cells}
\alias{format_cells}
\title{Format numeric table cells}
\usage{
format_cells(x, format.args = NULL)
}
\arguments{
\item{x}{data.frame or matrix.}

\item{format.args}{List. A named list of arguments to be passed to \code{\link{apa_num}} to format numeric values.}
}
\description{
Format all numeric cells of a table using \code{\link{apa_num}}.
\emph{This function is not exported.}
}
\seealso{
\code{\link[=apa_num]{apa_num()}}
}
\keyword{internal}
