% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{formula_processor}
\alias{formula_processor}
\title{Formula Interface for Plot Functions}
\usage{
formula_processor(formula, data, .fun, ...)
}
\arguments{
\item{formula}{A two-sided formula specifying the dependent variable on the left
of a \code{~} operator, and factors on the right. A vertical bar (\code{|}) can be used to
specify an additional subject identifier, which must be provided if an additional
aggregation step is necessary. See the examples section for details.}

\item{data}{A data frame (or list) from which the variables in \code{formula} should be taken.}
}
\description{
This is an internal function that ensures consistent processing of formulae
in plotting functions.
}
\keyword{internal}
