% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_df.R
\name{fuse_df}
\alias{fuse_df}
\alias{fuse_df.apa_results}
\alias{fuse_df.apa_results_table}
\title{Fuse Degrees-of-Freedom Columns into Variable Labels}
\usage{
fuse_df(x, quietly = FALSE, ...)

\method{fuse_df}{apa_results}(x, quietly = FALSE, ...)

\method{fuse_df}{apa_results_table}(x, quietly = FALSE, ...)
}
\arguments{
\item{x}{Either the complete output object created by \code{\link[=apa_print]{apa_print()}} methods,
or the \code{table} component of such objects.}

\item{quietly}{Logical. If FALSE (the default), checks if df columns contain non-integer values.}

\item{...}{further arguments passed from an to other methods}
}
\description{
Takes the output from \code{\link[=apa_print]{apa_print()}} methods and modifies the results table
by fusing information about degrees of freedom into the variable labels of
test-statistic columns.
}
\examples{
  apa_out <- apa_print(aov(yield ~ N * P, npk))

  # Standard output with separate columns for degrees of freedom:
  apa_out

  # Modified output where degrees of freedom are incorporated into the variable
  # label of column 'statistic':
  fuse_df(apa_out)


}
