% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_apa_results.R
\name{glue_apa_results}
\alias{glue_apa_results}
\alias{add_glue_to_apa_results}
\title{Create a new `apa_results` object}
\usage{
glue_apa_results(x = NULL, ...)

add_glue_to_apa_results(
  ...,
  est_glue,
  stat_glue,
  container,
  sublist = NULL,
  term_names = NULL,
  in_paren = FALSE
)
}
\arguments{
\item{x}{An environment, list or data frame used to look up values for
substitution.}

\item{...}{[\code{expressions}]\cr Unnamed arguments are taken to be expressions
string(s) to format. Multiple inputs are concatenated together before formatting.
Named arguments are taken to be temporary variables available for substitution.}

\item{est_glue}{Character. (Named vector of) expressions string(s) to
 format. Each string creates a new (named) element in the
`estimate`-sublist.}

\item{stat_glue}{Character. (Named vector of) expressions string(s) to
 format. Each string creates a new (named) element in the
`statistic`-sublist.}

\item{container}{List of class `apa_results` to add the glued results to.}

\item{sublist}{Character. Name of (new) sublist in `estimate`
`statistics`, and `full_result` to append glued results to (e.g.,
`modelfit`).}

\item{term_names}{Character. Used as names for the `estimate`-,
`statistics`-, and `full_result`-sublists, if multiple estimates or
statistics are glued.}

\item{in_paren}{Logical. Whether the formatted string is to be reported in
parentheses. If `TRUE`, parentheses in the formatted string (e.g., those
enclosing degrees of freedom) are replaced with brackets.}
}
\value{
Returns a list of class `apa_results`
}
\description{
Typeset the contents of an object according to the specified expression
strings and create a new or extend an existing `apa_results` object.
}
\examples{
# Tidy and typeset output
iris_lm <- lm(Sepal.Length ~ Petal.Length + Petal.Width, iris)
tidy_iris_lm <- broom::tidy(iris_lm, conf.int = TRUE)
tidy_iris_lm$p.value <- printp(tidy_iris_lm$p.value)

glance_iris_lm <- broom::glance(iris_lm)
glance_iris_lm$p.value <- printp(glance_iris_lm$p.value, add_equals = TRUE)
glance_iris_lm$df <- printnum(as.integer(glance_iris_lm$df))
glance_iris_lm$df.residual <- printnum(as.integer(glance_iris_lm$df.residual))

# Create `apa_results`-list
lm_results <- glue_apa_results(
    x = tidy_iris_lm
    , df = glance_iris_lm$df.residual
    , est_glue = "$b = <<estimate>>, 95\% CI $[<<conf.low>>,~<<conf.high>>]$"
    , stat_glue = "$t(<<df>>) = <<statistic>>$, $p <<p.value>>$"
    , term_names = make.names(names(coef(iris_lm)))
)

# Add modelfit information
add_glue_to_apa_results(
    .x = glance_iris_lm
    , container = lm_results
    , sublist = "modelfit"
    , est_glue = c(
        r2 = "$R^2 = <<r.squared>>$"
        , aic = ""
    )
    , stat_glue = c(
        r2 = "$F(<<df>>, <<df.residual>>) = <<statistic>>$, $p <<papaja:::add_equals(p.value)>>$"
        , aic = "$\\\\mathrm{AIC} = <<AIC>>$"
    )
)
}
