% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{hd_int}
\alias{hd_int}
\title{Highest density interval}
\usage{
hd_int(x, level = 0.95)
}
\arguments{
\item{x}{Numeric. A vector of observations.}

\item{level}{Numeric. Defines the width of the interval. Defaults to 95\% highest density intervals.}
}
\description{
Calculates the highest density interval of a vector of values
}
