% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{indent_stubs}
\alias{indent_stubs}
\title{Add stub indentation}
\usage{
indent_stubs(x, lines, filler = "\\\\ \\\\ \\\\ ")
}
\arguments{
\item{x}{data.frame.}

\item{lines}{List. A named list of vectors of length 2 giving the first and second row to
indent. Names of list elements will be used as titles for indented sections.}

\item{filler}{Character. Symbols used to indent stubs.}
}
\description{
Indents stubs by line and adds section headings
\emph{This function is not exported.}
}
\examples{
NULL
}
\seealso{
\code{\link{apa_table}}
}

