% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{initialize,annotated_vector-method}
\alias{initialize,annotated_vector-method}
\title{Generate an Annotated Vector}
\usage{
\S4method{initialize}{annotated_vector}(.Object, ...)
}
\description{
\code{\link{initialize}}-method for class union \code{annotated_vector}.
It takes care that \code{.Object@annotation@label} is copied to
\code{.Object@label}.
}
\keyword{internal}
