% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\docType{methods}
\name{label,vector-method}
\alias{label,vector-method}
\title{Extract label of an atomic vector}
\usage{
\S4method{label}{vector}(object)
}
\description{
This is only intended as a dummy method. However, it could be used to obtain compatibility with S3-type labels (like in Hmisc package).
}
\keyword{internal}
