% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\docType{methods}
\name{label}
\alias{label}
\alias{label,annotated_vector-method}
\alias{label,data.frame-method}
\alias{label<-,annotated_vector,character-method}
\alias{label<-,vector,character-method}
\alias{label<-,factor,character-method}
\alias{label<-,data.frame,character-method}
\title{Generic functions for 'label'}
\usage{
label(object)

\S4method{label}{annotated_vector}(object)

\S4method{label}{data.frame}(object)

\S4method{label}{annotated_vector,character}(object) <- value

\S4method{label}{vector,character}(object) <- value

\S4method{label}{factor,character}(object) <- value

\S4method{label}{data.frame,character}(object) <- value
}
\arguments{
\item{object}{blablabla}

\item{value}{bla}
}
\description{
What it does, more specifically...
}
