% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{levels<-,annotated_factor-method}
\alias{levels<-,annotated_factor-method}
\title{Set Levels of an Annotated Factor}
\usage{
\S4method{levels}{annotated_factor}(x) <- value
}
\arguments{
\item{x}{An object of \code{annotated_factor-class}.}

\item{value}{Character. The valid value for \code{levels(x)}.}
}
\description{
This method is as wrapper around \code{\link{levels}} and should do the same.
Annotations are preserved.
}
\keyword{internal}
