% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_plots.R
\name{lines.matrix}
\alias{lines.matrix}
\title{Matrix method for lines}
\usage{
\method{lines}{matrix}(x, y, type = "l", ...)
}
\description{
Internal function for convenient plotting of multiple lines.
}
\keyword{internal}
