% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{localize}
\alias{localize}
\title{Define Phrases According to Locale}
\usage{
localize(x)
}
\arguments{
\item{x}{Integer. Locale.}
}
\description{
Internal function. Defines phrases used throughout the manuscript.
}
\keyword{internal}
