% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\name{lookup_names}
\alias{lookup_names}
\alias{lookup_labels}
\alias{lookup_adjust_names}
\title{Lookup Tables for Column Names and Variable Labels}
\usage{
lookup_names

lookup_labels

lookup_adjust_names(x)
}
\description{
\code{apa_print()} converts many statistical output objects to output objects that
contain an \code{apa_results_table/data.frame} with standardized column names and
variable labels. For this purpose, it relies on the lookup tables provided
here.

\code{apa_print()} converts many statistical output objects that include
inferential statistics adjusted for multiple comparisons. To make these
adjustments transparent the statistics get an index with the
corresponding name. This function returns the proper names for these indices.
}
\keyword{internal}
