% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{names<-,annotated_vector,ANY-method}
\alias{names<-,annotated_vector,ANY-method}
\alias{names<-,annotated_named_vector,ANY-method}
\title{The Names of an Annotated Vector}
\usage{
\S4method{names}{annotated_vector,ANY}(x) <- value

\S4method{names}{annotated_named_vector,ANY}(x) <- value
}
\arguments{
\item{x}{An annotated vector.}

\item{value}{A character vector of the same length as x, or NULL.}
}
\description{
Methods to set the names of an annotated vector.
If names are specified for an unnamed annotated_vector, its class is changed
to the corresponding class from annotated_named_vector-superclass.
}
\details{
See \code{\link{names}} for details. If the names of an object of an
  annotated_named_vector are removed by setting \code{names(x) <- NULL}, it
  is coerced to the corresponding class from annotated_vector-superclasses
  (i.e., an annotated vector without a names slot).
}
\keyword{internal}
