% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_available}
\alias{package_available}
\title{Package available}
\usage{
package_available(x)
}
\arguments{
\item{x}{Character. Name of the package to be checked.}
}
\value{
Logical. Is the specified package installed?
}
\description{
Internal function to check if a specified package is installed.
}
\keyword{internal}
