% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_bcp47}
\alias{parse_bcp47}
\title{Simple language names from BCP 47 tags}
\usage{
parse_bcp47(x)
}
\arguments{
\item{x}{Character. BCP 47 tag.}
}
\value{
Character. Simple language name.
}
\description{
Internal function to translate BCP 47 tags to simple language names.
}
\keyword{internal}
