% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_ellipsis}
\alias{parse_ellipsis}
\title{Create variables from ellipsis}
\usage{
parse_ellipsis(...)
}
\arguments{
\item{...}{Variable names and values to create in the calling environment.}
}
\description{
Variables passed to a function using the ellipsis are assigned to the calling environment according to their
list names. \emph{This function is not exported.}
}
\examples{
parse_ellipsis(a = 10, b = 10)
a
b
}

