% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{points.matrix}
\alias{points.matrix}
\title{Matrix Method for points()}
\usage{
\method{points}{matrix}(x, y, type = "p", ...)
}
\description{
Internal function for convenient plotting of multiple points.
\emph{This function is not exported.}
}
\keyword{internal}
