% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_anova.R
\name{print_anova}
\alias{print_anova}
\title{Format statistics from ANOVA (APA 6th edition)}
\usage{
print_anova(x, observed = NULL, es = "ges", in_paren = FALSE)
}
\arguments{
\item{x}{Data.frame. A \code{data.frame} of class \code{apa_variance_table} as returned by \code{\link{arrange_anova}}.}

\item{observed}{Character. The names of the factors that are observed, (i.e., not manipulated). Necessary for calculation of generalized eta-squared; otherwise ignored.}

\item{es}{Character. The effect-size measure to be calculated; can be either \code{ges} for generalized eta-squared or \code{pes} for partial eta-squared.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}
}
\value{
A named list containing the following components:

   \describe{
     \item{\code{stat}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{est}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A named list of character strings comprised of \code{est} and \code{stat} for each factor.}
     \item{\code{table}}{A data.frame containing the complete ANOVA table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
This function is the workhorse of the \code{apa_print}-family for ANOVA. It takes a \code{data.frame}
of class \code{apa_variance_table} and produces strings to report the results in accordance with APA manuscript
guidelines. \emph{This function is not exported.}
}
\examples{
## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   anova_table <- arrange_anova(summary(npk_aov))
   print_anova(anova_table)
}
\seealso{
\code{\link{arrange_anova}}, \code{\link{apa_print.aov}}
}

