% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_confint}
\alias{print_confint}
\title{Create confidence interval string}
\usage{
print_confint(x, conf_level = NULL, ...)
}
\arguments{
\item{x}{Numeric. Either a \code{vector} of length 2 with attribute \code{conf.level} or a two-column \code{matrix}
and confidence region bounds as column names (e.g. "2.5 \%" and "97.5 \%") and coefficient names as row names.}

\item{conf_level}{Numeric. Vector of length 2 giving the lower and upper bounds of the confidence region in case
they cannot be determined from column names or attributes of \code{x}.}

\item{...}{Arguments to pass to \code{\link{printnum}}.}
}
\description{
Creates a character string from an object with attribute. \emph{This function is not exported.}
}
\examples{
\dontrun{
print_confint(c(1, 2), conf_level = 0.95)
}
}
\seealso{
\code{\link{printnum}}
}

