% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printnum.R
\name{print_df}
\alias{print_df}
\title{Print Degrees of Freedom}
\usage{
print_df(x, digits = 2L)
}
\description{
This is an internal function for processing degrees of freedom. It takes care
that trailing digits are only printed if non-integer values are given.
}
\keyword{internal}
