% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interval.R
\name{print_interval}
\alias{print_interval}
\alias{print_interval.data.frame}
\alias{print_interval.numeric}
\alias{print_interval.matrix}
\alias{print_confint}
\alias{print_hdint}
\title{Typeset Interval Estimate}
\usage{
print_interval(
  x,
  ...,
  conf_level = NULL,
  interval_type = NULL,
  enclose_math = FALSE
)

\method{print_interval}{data.frame}(x, ...)

\method{print_interval}{numeric}(
  x,
  y = NULL,
  conf_level = NULL,
  interval_type = NULL,
  enclose_math = FALSE,
  ...
)

\method{print_interval}{matrix}(
  x,
  conf_level = NULL,
  interval_type = NULL,
  enclose_math = FALSE,
  ...
)

print_confint(x, ..., interval_type = "CI")

print_hdint(x, ..., interval_type = "HDI")
}
\arguments{
\item{x}{Numeric. A vector (of length 2, unless \code{y} is also specified) with,
a two-column \code{matrix}, or a \code{data.frame}, which can coerced to a \code{matrix}.}

\item{...}{
  Arguments passed on to \code{\link[=printnum]{printnum}}
  \describe{
    \item{\code{}}{}
  }}

\item{conf_level}{Numeric. Confidence level of the interval. Ignored if
level can be infered from attributes of \code{x}, see Details.}

\item{interval_type}{Character. Abbreviation indicating the type of interval
estimate, e.g. \code{CI}.}

\item{enclose_math}{Logical. Indicates whether the interval should be
enclosed in \code{$} (i.e., a math environment).}

\item{y}{Numeric. An optional vector of the same length as \code{x}.}
}
\value{
A singel interval is returned as a \code{character} vector of length 1;
multiple intervals are returned as a named \code{list} of \code{character} vectors
of length 1.
}
\description{
Creates a character string to report interval estimates, such as frequentist
confidence or Bayesian credible intervals.
}
\details{
If possible the confidence level of the interval is inferred from
attributes of \code{x}. For a vector of length 2, the attribute \code{conf.level} is
is consulted; for a \code{matrix} or \code{data.frame} the column names are used,
if they are of the format "2.5 \\%" and "97.5 \\%".

If \code{x} is a \code{matrix} or \code{data.frame} the row names are used as names for
the returned \code{list} of intervals.
}
\examples{
\dontrun{
print_confint(1, 2, conf_level = 0.95)
print_confint(c(1, 2), conf_level = 0.95)
print_confint(matrix(c(1, 2), ncol = 2), conf_level = 0.95)
print_confint(confint(lm(cars)))
print_confint(confint(lm(cars)), digits = 3)
}
}
\seealso{
\code{\link{printnum}}
}
\keyword{internal}
