% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_comp.R
\name{print_model_comp}
\alias{print_model_comp}
\title{Typeset Statistical Results from Model Comparisons}
\usage{
print_model_comp(
  x,
  models = NULL,
  conf.int = NULL,
  boot_samples = 1000,
  progress_bar = FALSE,
  in_paren = FALSE,
  observed = TRUE
)
}
\arguments{
\item{x}{A data frame of class \code{apa_variance_table} as returned by \code{\link[=arrange_anova]{arrange_anova()}}.}

\item{models}{List. List containing fitted \code{lm} objects that were compared using \code{\link[=anova]{anova()}}. If the list is named, element names are used as model names in the output object.}

\item{conf.int}{Numeric. Confidence level for the confidence interval for \eqn{\Delta R^2} if \code{x} is a model comparison object of class \code{anova}. If \code{conf.int = NULL} no confidence intervals are estimated.}

\item{boot_samples}{Numeric. Number of bootstrap samples to estimate confidence intervals for \eqn{\Delta R^2} if \code{x} is a model comparison object of class \code{anova}; ignored if \code{conf.int = NULL}.}

\item{progress_bar}{Logical. Determines whether a progress bar is printed while bootstrapping.}

\item{in_paren}{Logical. Whether the formatted string is to be reported in
parentheses. If \code{TRUE}, parentheses in the formatted string (e.g., those
enclosing degrees of freedom) are replaced with brackets.}

\item{observed}{Logical. Indicates whether predictor variables were
observed. See details.}
}
\value{
A named list (with additional class \code{apa_results}) containing the following components:

\describe{
\item{\code{statistic}}{
A named list of character strings giving the test statistic, parameters, and \emph{p} value for each factor.
}
\item{\code{estimate}}{
A named list of character strings giving the effect size estimates for each factor,
either in units of the analyzed scale or as standardized effect size.
}
\item{\code{full_result}}{
A named list of character strings comprised of \code{estimate} and \code{statistic} for each factor.
}
\item{\code{table}}{
A data.frame containing the complete comparison table, which can be passed to \code{\link[=apa_table]{apa_table()}}.
}
}
}
\description{
This function is the workhorse of the \link[=apa_print.list]{apa_print()} method
for model comparisons. It takes a data frame of class \code{apa_model_comp} and
produces strings to report the results in accordance with APA manuscript
guidelines.
\emph{This function is not exported.}
}
\examples{
 \dontrun{
   mod1 <- lm(Sepal.Length ~ Sepal.Width, data = iris)
   mod2 <- update(mod1, formula = . ~ . + Petal.Length)
   mod3 <- update(mod2, formula = . ~ . + Petal.Width)

   # No bootstrapped Delta R^2 CI
   print_model_comp(list(Baseline = mod1, Length = mod2, Both = mod3), boot_samples = 0)
 }
}
\seealso{
\code{\link[=arrange_anova]{arrange_anova()}}, \code{\link[=apa_print.aov]{apa_print.aov()}}
}
\keyword{internal}
