% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printnum.R
\name{printnum}
\alias{printnum}
\alias{printp}
\title{Prepare numeric values for printing}
\usage{
printnum(x, digits = 2, gt1 = TRUE, zero = TRUE, margin = 1)

printp(x, margin = 1)
}
\arguments{
\item{x}{Value to format; can be either a single value, vector, or matrix.}

\item{digits}{Integer. Number of decimal places to print.}

\item{gt1}{Logical. Indicates if the absolute value of the statistic can, in principal, exceed 1.}

\item{zero}{Logical. Indicates if the statistic can, in principal, be 0.}

\item{margin}{Integer. If \code{x} is a matrix, the function is applied either across rows (= 1)
or columns (= 2).}
}
\description{
This function formats numeric values according to APA guidelines (6th edition) for reporting.
}
\details{
If \code{x} is a vector, \code{digits}, \code{gt1}, and \code{zero} can be vectors
   according to which each element of the vector is formated. Parameters are recycled if length of \code{x}
   exceeds length of the parameter vectors. If \code{x} is a matrix, the vectors specify the formating
   of either rows or columns according to the value of \code{margin}.
}
\section{Functions}{
\itemize{
\item \code{printp}: Convenience wrapper for \code{printnum} to print p-values.
}}
\examples{
...
}
\seealso{
\code{\link{apply}}
}

