% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{rep,annotated_vector-method}
\alias{rep,annotated_vector-method}
\alias{rep,annotated_factor-method}
\title{Replicate Elements of Annotated Vectors}
\usage{
\S4method{rep}{annotated_vector}(x, ...)

\S4method{rep}{annotated_factor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{annotated_vector}.}

\item{...}{Further arguments that can be passed to\code{\link{rep}}.}
}
\description{
Methods for annotated vectors to replicate values, calls \code{\link{rep}}
and preserves annotations.
}
\keyword{internal}
