% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{names<-,annotated_vector,NULL-method}
\alias{names<-,annotated_vector,NULL-method}
\alias{names<-,annotated_named_vector,NULL-method}
\title{Replace the Names Slot with NULL}
\usage{
\S4method{names}{annotated_vector,`NULL`}(x) <- value

\S4method{names}{annotated_named_vector,`NULL`}(x) <- value
}
\description{
If \code{value(x) <- NULL} is assigned, an unnamed vector keeps its class,
a named vector is coerced to its unnamed complement.
}
\keyword{internal}
