% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{round_cells}
\alias{round_cells}
\title{Round table cells}
\usage{
round_cells(x, digits = getOption("digits"),
  na_string = getOption("papaja.na_string"))
}
\arguments{
\item{x}{data.frame or matrix.}

\item{digits}{Numeric. Vector of the number of digits to round numeric columns to.}

\item{na_string}{Character. String to print if element of \code{x} is \code{NA}.}
}
\description{
Round all numeric cells of a table using \code{\link{printnum}}.
\emph{This function is not exported.}
}
\examples{
NULL
}
\seealso{
\code{\link{printnum}}
}

