% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{set_paren}
\alias{set_paren}
\title{Set parentheses or brackets}
\usage{
set_paren(in_paren = FALSE)
}
\arguments{
\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses.}
}
\description{
Define \code{op} and \code{cp} to generate strings in the calling environment. \emph{This function is not exported.}
}
\examples{
set_op_cp()
}

