% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{show,annotated_vector-method}
\alias{show,annotated_vector-method}
\alias{show,annotated_named_vector-method}
\alias{show,annotated_factor-method}
\alias{show,annotated_named_factor-method}
\title{Show an Annotated Vector}
\usage{
\S4method{show}{annotated_vector}(object)

\S4method{show}{annotated_named_vector}(object)

\S4method{show}{annotated_factor}(object)

\S4method{show}{annotated_named_factor}(object)
}
\description{
Display an annotated vector by printing to the console.
}
\keyword{internal}
