% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{show,vector_annotation-method}
\alias{show,vector_annotation-method}
\title{Show a vector annotation}
\usage{
\S4method{show}{vector_annotation}(object)
}
\description{
Display a vector annotation by printing to the console.
}
\keyword{internal}
