% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_columns}
\alias{sort_columns}
\title{Sort the Columns of an APA Results Table}
\usage{
sort_columns(x)
}
\arguments{
\item{x}{A \code{data.frame} with standardized column names.}
}
\description{
An internal function that sorts the columns of a `data.frame` according to
our standards.
}
\keyword{internal}
