% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_effects}
\alias{sort_effects}
\title{Sort ANOVA table by effects}
\usage{
sort_effects(x)
}
\arguments{
\item{x}{data.frame. An arbitrary data.frame with a column named "Effect", e.g., a table element
produced by \link{\code{apa_print}}.}
}
\value{
Returns the same data.frame with reordered rows.
}
\description{
Sorts rows in ANOVA table produced by \link{\code{apa_print}} by complexity (i.e., main effects,
two-way interactions, three-way interactions, etc.).
}
\examples{
## From Venables and Ripley (2002) p. 165.
npk_aov <- aov(yield ~ block + N * P * K, npk)
npk_aov_results <- apa_print(npk_aov)
sort_effects(npk_aov_results$table)
}

