% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{svl}
\alias{svl}
\alias{strip_math_tags}
\title{Strip Math Tags from Variable Labels or Strings}
\usage{
svl(x, use_math = FALSE)

strip_math_tags(x)
}
\arguments{
\item{x}{A (labelled) character string.}
}
\description{
Internal function to strip math tags from variable labels or strings. \code{svl()}
returns the stripped variable label of \code{x}, if available. \code{strip_math_tags} returns
the stripped character \code{x}.
}
\keyword{internal}
