% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{methods}
\name{[,annotated_vector,ANY,ANY,ANY-method}
\alias{[,annotated_vector,ANY,ANY,ANY-method}
\alias{[,annotated_factor,ANY,ANY,ANY-method}
\title{Indexing of annotated vectors}
\usage{
\S4method{[}{annotated_vector,ANY,ANY,ANY}(x, i)

\S4method{[}{annotated_factor,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{The annotated vector object to be subsetted.}

\item{i}{An}
}
\description{
When subsetting a vector by "[", meta-data (e.g., label and unit of measurement)
are preserved.
}
\examples{
a <- new("annotated_numeric", .Data = 1:4)
test <- a[1:2]
}
\keyword{internal}
