% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{summary.papaja_wsci}
\alias{summary.papaja_wsci}
\title{Summarise Within-Subjects Confidence Intervals}
\usage{
\method{summary}{papaja_wsci}(object, ...)
}
\arguments{
\item{object}{An object of class \code{papaja_wsci}, generated with function \code{\link[=wsci]{wsci()}}.}

\item{...}{Further arguments that may be passed, currently ignored.}
}
\description{
Calculate upper and lower limits of within-subjects confidence intervals calculated
with \code{\link[=wsci]{wsci()}} and return them along their respective means.
}
