% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate}
\alias{validate}
\title{Validate function input}
\usage{
validate(x, name = NULL, check_class = NULL, check_mode = NULL,
  check_integer = FALSE, check_NA = TRUE, check_infinite = TRUE,
  check_length = NULL, check_dim = NULL, check_range = NULL)
}
\arguments{
\item{x}{Function input.}

\item{name}{Character. Name of variable to validate; if \code{NULL} variable name of object supplied to \code{x} is used.}

\item{check_class}{Character. Name of class to expect.}

\item{check_mode}{Character. Name of mode to expect.}

\item{check_integer}{Logical. If \code{TRUE} an object of type \code{integer} or a whole number \code{numeric} is expected.}

\item{check_NA}{Logical. If \code{TRUE} an non-\code{NA} object is expected.}

\item{check_infinite}{Logical. If \code{TRUE} a finite object is expected.}

\item{check_length}{Integer. Length of the object to expect.}

\item{check_dim}{Numeric. Vector of object dimensions to expect.}

\item{check_range}{Numeric. Vector of length 2 defining the expected range of the object.}
}
\description{
This function can be used to validate the input to functions. \emph{This function is not exported.}
}
\examples{
\dontrun{
in_paren <- TRUE # Taken from printnum()
validate(in_paren, check_class = "logical", check_length = 1)
validate(in_paren, check_class = "numeric", check_length = 1)
}
}

