% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\docType{methods}
\name{variable_label,labelled-method}
\alias{variable_label,labelled-method}
\title{Extract label from a 'labelled' vector}
\usage{
\S4method{variable_label}{labelled}(object)
}
\arguments{
\item{object}{A vector of (S3-) class \code{labelled} (e.g., from \pkg{Hmisc} package) and attribute \code{label} set.}
}
\description{
This method is for minimum compatibility with the \pkg{Hmisc} package. Allows to extract a variable label that was set by \code{Hmisc::label}.
}
