% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\docType{methods}
\name{variable_label,vector-method}
\alias{variable_label,vector-method}
\title{Extract label of an Atomic (i.e., non-annotated) vector}
\usage{
\S4method{variable_label}{vector}(object)
}
\description{
This is only intended as a dummy method. However, it could be used to obtain compatibility with S3-type labels (like in Hmisc package).
}
\keyword{internal}
