% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{assign_label}
\alias{assign_label}
\alias{assign_label.default}
\alias{assign_label.data.frame}
\alias{variable_label}
\alias{variable_label.default}
\alias{variable_label.data.frame}
\alias{variable_label<-}
\alias{variable_label<-.default}
\alias{variable_label<-.data.frame}
\title{Variable labels}
\usage{
assign_label(x, value)

\method{assign_label}{default}(x, value)

\method{assign_label}{data.frame}(x, value)

variable_label(x)

\method{variable_label}{default}(x)

\method{variable_label}{data.frame}(x)

variable_label(x) <- value

\method{variable_label}{default}(x) <- value

\method{variable_label}{data.frame}(x) <- value
}
\arguments{
\item{x}{Either a vector or a \code{data.frame} containing the variables you want to label.}

\item{value}{A vector with the variable label(s) to be assigned. Can be anything, e.g. character, expression, etc.}
}
\value{
\code{variable_label} returns the variable labels stored as attributes to the vector or the columns of a \code{data.frame}.
        \code{assign_label} returns \code{x} with the added variable labels.
}
\description{
Functions used to assign and extract variable labels of a vector or the columns (i.e., vectors) of a \code{data.frame}.
}
