% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{variable_unit}
\alias{variable_unit}
\alias{variable_unit.default}
\alias{variable_unit.data.frame}
\alias{variable_unit<-}
\alias{variable_unit<-.default}
\alias{variable_unit<-.data.frame}
\title{Assign and extract variable units of measurement}
\usage{
variable_unit(x)

\method{variable_unit}{default}(x)

\method{variable_unit}{data.frame}(x)

variable_unit(x) <- value

\method{variable_unit}{default}(x) <- value

\method{variable_unit}{data.frame}(x) <- value
}
\arguments{
\item{x}{Either a vector or a \code{data.frame} containing the variables you want to assign a unit of measurement.}

\item{value}{A vector with the variable unit(s) to be assigned. Can be anything, e.g. character, expression, etc.}
}
\value{
If applied to a \code{data.frame}, \code{variable_label} returns a named vector containing all variable units.
        If applied to a \code{vector}, a length-one \code{character}.
}
\description{
Functions used to assign and extract variable units (of measurement) of a \code{vector} \emph{or}
the columns (i.e., vectors) of a \code{data.frame}.
}
\examples{
variable_unit(npk$yield) <- "pounds/plot"
variable_unit(npk)
}
