% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{class}
\name{vector_annotation-class}
\alias{vector_annotation-class}
\title{Vector annotations}
\description{
A class intended to keep all meta-data of a vector.
}
\section{Slots}{

\describe{
\item{\code{label}}{Length-one character. A label for the vector.}

\item{\code{unit}}{Length-one character. The unit of measurement.}
}}

\keyword{internal}
