% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_title_page.R
\name{word_title_page}
\alias{word_title_page}
\title{Create title page}
\usage{
word_title_page(x)
}
\arguments{
\item{x}{List. Meta data of the document as a result from \code{\link[yaml]{yaml.load}}.}
}
\description{
Creates text for the title and abstract page for MS Word documents.
\emph{This function is not exported.}
}
\seealso{
\code{\link{apa6_word}}
}
\keyword{internal}
