% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa6_formats.R
\name{apa6_pdf}
\alias{apa6_pdf}
\alias{apa6_word}
\title{APA article (6th edition)}
\usage{
apa6_pdf(class = "man", fig_caption = TRUE, pandoc_args = NULL,
  keep.tex = TRUE, ...)

apa6_word(class = "man", fig_caption = TRUE, pandoc_args = NULL, ...)
}
\arguments{
\item{class}{Character. Specifies if the document will be a manuscript ("man"), a document ("doc"),
or a fully typsetted journal article ("jou"). Currently only manuscripts are supported for
MS Word documents.}

\item{fig_caption}{Logical. Indicates if figures are rendered with captions.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Further arguments to pass to \code{rmarkdown::pdf_document} or \code{rmarkdown::word_document}.}

\item{Logical.}{Keep the intermediate tex file used in the conversion to PDF.}
}
\description{
Template for creating an article according to APA guidelines (6th edition) in PDF format.
}
\section{Functions}{
\itemize{
\item \code{apa6_word}: Format to create .docx-files. \code{class} parameter is ignored.
}}
\examples{
NULL
}

