% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa6_formats.R
\name{apa6_pdf}
\alias{apa6_pdf}
\alias{apa6_docx}
\alias{apa6_word}
\title{APA article (6th edition)}
\usage{
apa6_pdf(fig_caption = TRUE, number_sections = FALSE, toc = FALSE,
  keep_tex = TRUE, md_extensions = NULL, ...)

apa6_docx(fig_caption = TRUE, pandoc_args = NULL,
  md_extensions = NULL, ...)

apa6_word(fig_caption = TRUE, pandoc_args = NULL,
  md_extensions = NULL, ...)
}
\arguments{
\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{number_sections}{\code{TRUE} to number section headings}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for additional details.}

\item{...}{Further arguments to pass to \code{\link[rmarkdown]{pdf_document}} or \code{\link[rmarkdown]{word_document}}.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Template for creating an article according to APA guidelines (6th edition) in PDF format.
}
\details{
When creating PDF documents the YAML option \code{classoption} is passed to the class options of the LaTeX apa6 document class.
   In this case, additional options are available. Refer to the apa6 document class
   \href{ftp://ftp.fu-berlin.de/tex/CTAN/macros/latex/contrib/apa6/apa6.pdf}{documentation} to find out about class options
   such as paper size or draft watermarks.

   When creating PDF documents the output device for figures defaults to \code{c("pdf", "png")},
   so that each figure is saved in all four formats at a resolution of 300 dpi.
}
\section{Functions}{
\itemize{
\item \code{apa6_docx}: Format to create .docx-files. \code{class} parameter is ignored. \emph{This function
should be considered experimental.}

\item \code{apa6_word}: Format to create .docx-files. Alias of \code{apa6_docx}.
}}

\examples{
NULL
}
\seealso{
\code{\link[bookdown]{html_document2}}, \code{\link[rmarkdown]{pdf_document}}, \code{\link[rmarkdown]{word_document}}
}
