% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa6_formats.R
\name{apa6_pdf}
\alias{apa6_pdf}
\alias{apa6_word}
\title{APA article (6th edition)}
\usage{
apa6_pdf(class = "man", fig_caption = TRUE, pandoc_args = NULL,
  keep_tex = TRUE, ...)

apa6_word(class = "man", fig_caption = TRUE, pandoc_args = NULL, ...)
}
\arguments{
\item{class}{Character. Specifies if the document will be a manuscript (\code{man}), a document (\code{doc}),
or a fully typsetted journal article (\code{jou}). Currently only manuscripts are supported for
MS Word documents. At this point coversion to MS word is considered experimental. See details.}

\item{fig_caption}{Logical. Indicates if figures are rendered with captions.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{keep_tex}{Logical. Keep the intermediate tex file used in the conversion to PDF.}

\item{...}{Further arguments to pass to \code{\link[rmarkdown]{pdf_document}} or \code{\link[rmarkdown]{word_document}}.}
}
\description{
Template for creating an article according to APA guidelines (6th edition) in PDF format.
}
\details{
When creating PDF documents \code{class} is passed to the class options of the LaTeX apa6 document class. In this case,
   additional options are available. Refer to the apa6 document class
   \href{ftp://ftp.fu-berlin.de/tex/CTAN/macros/latex/contrib/apa6/apa6.pdf}{documentation} to find out about class options
   such as paper size or draft watermarks.

   When creating PDF documents the output device for figures defaults to \code{c("pdf", "postscript", "png", "tiff")},
   so that each figure is saved in all four formats at a resolution of 300 dpi.
}
\section{Functions}{
\itemize{
\item \code{apa6_word}: Format to create .docx-files. \code{class} parameter is ignored. \emph{This function
should be considered experimental.}
}}
\examples{
NULL
}
\seealso{
\code{\link[rmarkdown]{pdf_document}}, \code{\link[rmarkdown]{word_document}}
}

