% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_barplot.R
\name{apa_barplot}
\alias{apa_barplot}
\title{APA-compatible barplot}
\usage{
apa_barplot(data, id, factors, dv, tendency = mean, dispersion = conf_int,
  level = 0.95, fun_aggregate = mean, na.rm = TRUE, ylim = NULL,
  main = NULL, xlab = NULL, ylab = NULL, intercepts = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} that contains your data}

\item{id}{Character. Variable name that identifies subjects.}

\item{factors}{Character. A vector of variable names that shall be used to stratify data.}

\item{dv}{Character. The dependent variable.}

\item{tendency}{Closure. A function that will be used as measure of central tendency.}

\item{dispersion}{Closure. A function that will be used to constrct error bars (i.e., whiskers). Defaults to \code{conf_int} for 95\% confidence intervals. See details.}

\item{level}{Numeric. If confidence intervals are printed, this determines the area of the cdf that will be covered. Defaults to .95 for 95\% confidence intervals. If \code{level} is specified, confidence intervals are printed and \code{dispersion} is ignored.}

\item{fun_aggregate}{Closure. The function that will be used to aggregate data. Defaults to \code{mean}}

\item{na.rm}{Logical. Specifies if missing values will be removed. Defaults to \code{TRUE}.}

\item{ylim}{Numeric. A vector specifying upper and lower limit of the y-axis.}

\item{main}{Character.The main title for your plot.}

\item{xlab}{Character. Specifies the label of your x-axis. Defauls to the first element of \code{factors}.}

\item{ylab}{Character. Specifies the label of your y-axis. Defaults to \code{dv}.}

\item{intercepts}{Numeric. Draws horizontal lines into your plot. Specify one or multiple y-values.}

\item{...}{Further arguments than can be passed to the underlying \code{barplot} function.}
}
\description{
This function returns barplots.
}
\details{
The measure of dispersion can be either \code{conf_int} for confidence intervals, \code{se} for standard errors,
   or any other standard function. If \code{conf_int} is specified, you can also specify the area of the cumulative distribution function that will be covered.
   For instance, if you want a 98\% confindence interval, specify \code{level=.98}. \code{level} defaults to .95.
}
\examples{
NULL
}

