% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_beeplot.R
\name{apa_beeplot}
\alias{apa_beeplot}
\title{Beeswarm plots for factorial designs that conform to APA guidelines}
\usage{
apa_beeplot(data, id, factors, dv, tendency = mean, dispersion = conf_int,
  level = 0.95, fun_aggregate = mean, na.rm = TRUE, intercept = NULL,
  args_axis = list(), args_points = list(), args_lines = list(),
  args_arrows = list(), args_legend = list(), ...)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data.}

\item{id}{Character. Variable name that identifies subjects.}

\item{factors}{Character. A vector of up to 4 variable names that is used to stratify the data.}

\item{dv}{Character. The name of the dependent variable.}

\item{tendency}{Closure. A function that will be used as measure of central tendency.}

\item{dispersion}{Closure. A function that will be used to construct error bars (i.e., whiskers). Defaults to
\code{conf_int} for 95\% confidence intervals. See details.}

\item{level}{Numeric. Defines the width of the interval if confidence intervals are plotted. Defaults to 0.95
for 95\% confidence intervals. Ignored if \code{dispersion} is not \code{conf_int}.}

\item{fun_aggregate}{Closure. The function that will be used to aggregate observations within subjects and factors
before calculating descriptive statistics for each cell of the design. Defaults to \code{mean}.}

\item{na.rm}{Logical. Specifies if missing values are removed. Defaults to \code{TRUE}.}

\item{intercept}{Numeric. Adds a horizontal line to the plot. Can be either a single value or a matrix. For the matrix
case, multiple lines are drawn, where the dimensions of the matrix determine the number of lines to be drawn.}

\item{args_axis}{An optional \code{list} that contains further arguments that may be passed to \code{\link{axis}}}

\item{args_points}{An optional \code{list} that contains further arguments that may be passed to \code{\link{points}}}

\item{args_lines}{An optional \code{list} that contains further arguments that may be passed to \code{\link{lines}}}

\item{args_arrows}{An optional \code{list} that contains further arguments that may be passed to \code{\link{arrows}}}

\item{args_legend}{An optional \code{list} that contains further arguments that may be passed to \code{\link{legend}}}

\item{...}{Further arguments than can be passed to \code{\link{plot}} function.}
}
\description{
Wrapper function that creates one or more lineplots from a data.frame containing data from
a factorial design and sets APA-friendly defaults. It sequentially calls \code{\link{plot}},
\code{\link{axis}}, \code{\link{points}}, \code{\link{lines}}, \code{\link{arrows}} and
\code{\link{legend}}, that may be further customized.
}
\details{
The measure of dispersion can be either \code{conf_int} for confidence intervals, \code{se} for standard errors,
   or any other standard function. If \code{conf_int} is specified, you can also specify the area of the cumulative
   distribution function that will be covered. For instance, if you want a 98\% confindence interval, specify
   \code{level = 0.98}. \code{level} defaults to 0.95.
}
\examples{
apa_beeplot(
   data = npk
   , id = "block"
   , dv = "yield"
   , factors = c("N")
)

apa_beeplot(
data = npk
 , id = "block"
 , dv = "yield"
 , factors = c("N", "P")
 , args.legend = list(x = "center")
 #, jit = 0
)

apa_beeplot(
   data = npk
   , id = "block"
   , dv = "yield"
   , factors = c("N", "P", "K")
   , ylim = c(0, 80)
   , level = .34
   , las = 1
)
}
\seealso{
\code{\link{barplot}}
}

