% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_factorial_plot.R
\name{apa_factorial_plot}
\alias{apa_factorial_plot}
\alias{apa_factorial_plot.default}
\alias{apa_factorial_plot.afex_aov}
\title{Plots for factorial designs that conform to APA guidelines}
\usage{
apa_factorial_plot(data, ...)

\method{apa_factorial_plot}{default}(data, id, factors = NULL, dv,
  tendency = mean, dispersion = conf_int, level = 0.95,
  fun_aggregate = mean, na.rm = TRUE, use = "all.obs",
  reference = 0, intercept = NULL, args_x_axis = NULL,
  args_y_axis = NULL, args_title = NULL, args_rect = NULL,
  args_points = NULL, args_lines = NULL, args_swarm = NULL,
  args_error_bars = NULL, args_legend = NULL, plot = NULL,
  jit = 0.3, xlab = NULL, ylab = NULL, main = NULL, ...)

\method{apa_factorial_plot}{afex_aov}(data, tendency = mean,
  dispersion = conf_int, fun_aggregate = mean, ...)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data or an object of class \code{afex_aov}}

\item{...}{Arguments passed on to \code{graphics::plot.window}
\describe{
  \item{xlim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}
  \item{ylim}{numeric vectors of length 2, giving the x and y
    coordinates ranges.}
  \item{log}{character; indicating which axes should be in log scale.}
  \item{asp}{numeric, giving the \bold{asp}ect ratio y/x,
    see \sQuote{Details}.}
}}

\item{id}{Character. Variable name that identifies subjects.}

\item{factors}{Character. A vector of up to four variable names that is used to stratify the data.}

\item{dv}{Character. The name of the dependent variable.}

\item{tendency}{Closure. A function that will be used as measure of central tendency.}

\item{dispersion}{Closure. A function that will be used to construct error bars (i.e., whiskers). Defaults to
\code{conf_int} for 95\% between-subjects confidence intervals. See details for more options, especially for within-subjects confidence intervals.}

\item{level}{Numeric. Defines the width of the interval if confidence intervals are plotted. Defaults to \code{0.95}.
for 95\% confidence intervals. Ignored if \code{dispersion} is not a confidence-interval function. See details.}

\item{fun_aggregate}{Closure. The function that will be used to aggregate observations within subjects and factors
before calculating descriptive statistics for each cell of the design. Defaults to \code{mean}.}

\item{na.rm}{Logical. Specifies if missing values are removed. Defaults to \code{TRUE}.}

\item{use}{Character. Specifies a method to exclude cases if there are missing values \emph{after} aggregating.
Possible options are \code{"all.obs"} or \code{"complete.obs"}.}

\item{reference}{Numeric. A reference point that determines the \emph{y} coordinate of the \emph{x} axis. Useful if there exists a 'nil' value; defaults to \code{0}.}

\item{intercept}{Numeric. Adds a horizontal line at height \code{intercept} to the plot. Can be either a single value or a matrix. For the matrix
case, multiple lines are drawn, where the dimensions of the matrix determine the number of lines to be drawn.}

\item{args_x_axis}{An optional \code{list} that contains further arguments that may be passed to \code{\link{axis}} for customising the \emph{x} axis.}

\item{args_y_axis}{An optional \code{list} that contains further arguments that may be passed to \code{\link{axis}} for customising the \emph{y} axis.}

\item{args_title}{An optional \code{list} that contains further arguments that may be passed to \code{\link{title}}.}

\item{args_rect}{An optional \code{list} that contains further arguments that may be passed to \code{\link{rect}}.}

\item{args_points}{An optional \code{list} that contains further arguments that may be passed to \code{\link{points}}.}

\item{args_lines}{An optional \code{list} that contains further arguments that may be passed to \code{\link{lines}}.}

\item{args_swarm}{An optional \code{list} that contains further arguments to customize the \code{\link{points}} of the beeswarm.}

\item{args_error_bars}{An optional \code{list} that contains further arguments that may be passed to \code{\link{arrows}}.}

\item{args_legend}{An optional \code{list} that contains further arguments that may be passed to \code{\link{legend}}}

\item{plot}{Character. A vector specifying which elements of the plot should be plotted. Available options are
\code{c("points", "error_bars", "bars", "swarms", "lines")}}

\item{jit}{Numeric. Determines the amount of horizontal displacement. Defaults to \code{0.3}, defaults to \code{0.4} if \code{plot = "bars"}.}

\item{xlab}{Character or expression. Label for \emph{x} axis.}

\item{ylab}{Character or expression. Label for \emph{y} axis.}

\item{main}{Character or expression. For up to two factors, simply specify the main title. If you stratify the data by more than two factors,
either specify a single value that will be added to automatically generated main title, \emph{or} specify an array of multiple titles, one for each plot area.}
}
\description{
Wrapper function that creates one or more plots by sequentially calling functions from the \pkg{graphics} package.
\code{apa_factorial_plot} is the workhorse function that is called by \code{\link{apa_barplot}}, \code{\link{apa_beeplot}}, and \code{\link{apa_lineplot}}.
}
\details{
The measure of dispersion can be either \code{conf_int} for between-subjects confidence intervals, \code{se} for standard errors,
   or any other standard function. For within-subjects confidence intervals, specify \code{wsci} or \code{within_subjects_conf_int}.

   If between- or within-subjects confidence intervals are requested, you can also specify the area of the cumulative
   distribution function that will be covered. For instance, if you want a 98\% confidence interval, specify
   \code{level = 0.98}. \code{level} defaults to 0.95.

\strong{Customisation of plot elements}

   \code{apa_factorial_plot} and its descendants \code{apa_barplot}, \code{apa_lineplot}, and \code{apa_beeplot} are wrapper functions that sequentially call
\code{\link{plot.new}},
\code{\link{plot.window}},
\code{\link{axis}} (once for \emph{x} axis, once for \emph{y} axis),
\code{\link{title}} for axis labels and titles,
\code{\link{rect}} for bars in barplots,
\code{\link{points}} for bee swarms,
\code{\link{lines}} for lines connecting central tendency points,
\code{\link{arrows}} for error bars,
\code{\link{points}} for tendency points,
\code{\link{legend}} for a legend, and
\code{\link{lines}} for intercepts.
These calls can be customised by setting the respective parameters \code{args_***}.
}
\examples{
apa_factorial_plot(
  data = npk
  , id = "block"
  , dv = "yield"
  , factors = c("N", "P", "K")
  , las = 1
  , plot = c("error_bars", "points", "swarms")
  , ylim = c(0, 100)
)
}
\seealso{
Other plots for factorial designs: \code{\link{apa_barplot}},
  \code{\link{apa_beeplot}}, \code{\link{apa_lineplot}}
}
\concept{plots for factorial designs}
