% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_prepare_doc.R
\name{apa_prepare_doc}
\alias{apa_prepare_doc}
\title{Prepare APA document}
\usage{
apa_prepare_doc()
}
\description{
Prepares objects to be used in the rendering process and creates title page and abstract for MS Word
documents.
}
\details{
The function creates and locks a non-exported object \code{apa_lang} that is used by other
   \pkg{papaja}-functions. \code{apa_lang} is a \code{list} containing localizations for document
   elements such as abstract and title. The selected language is defined by the \code{lang}-parameter
   in the documents yaml-header. Currently, English (default) and German ("german") are supported.

   If the output document is MS Word (\code{output: \link{apa6_word}}) the function additionally creates
   a title page and adds the abstract. You should, therefore, allways call \code{apa_prepare_doc} at the
   very beginning of the R Markdown document.
}
\examples{
apa_prepare_doc()
}
\seealso{
\code{\link{apa6_word}}
}

