% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_print_anova.R
\name{apa_print.aov}
\alias{apa_print.Anova.mlm}
\alias{apa_print.anova}
\alias{apa_print.aov}
\alias{apa_print.aovlist}
\alias{apa_print.summary.aov}
\alias{apa_print.summary.aovlist}
\title{Format statistics (APA 6th edition) from ANOVAs}
\usage{
\method{apa_print}{aov}(x, es = "ges", observed = NULL, in_paren = FALSE,
  ...)

\method{apa_print}{summary.aov}(x, es = "ges", observed = NULL,
  in_paren = FALSE, ...)

\method{apa_print}{aovlist}(x, es = "ges", observed = NULL,
  in_paren = FALSE, ...)

\method{apa_print}{summary.aovlist}(x, es = "ges", observed = NULL,
  in_paren = FALSE, ...)

\method{apa_print}{anova}(x, es = "ges", observed = NULL,
  in_paren = FALSE, ...)

\method{apa_print}{Anova.mlm}(x, correction = "GG", es = "pes",
  observed = NULL, in_paren = FALSE, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{es}{Character. The effect-size measure to be calculated; can be either \code{ges} for generalized eta-squared or \code{pes} for partial eta-squared.}

\item{observed}{Character. The names of the factors that are observed, (i.e., not manipulated). Is necessary for calculation of generalized eta-squared; otherwise ignored.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}

\item{...}{Additional arguments passed to or from other methods.}

\item{correction}{Character In the case of repeated-measures ANOVA, the type of sphericity correction to be used. Either \code{GG} for Greenhouse-Geisser or \code{HF} for Huyn-Feldt methods or \code{none} is also possible.}
}
\value{
\code{apa_print.aov} and related functions return a list containing the following components according to the input:

   \describe{
     \item{\code{stat}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{est}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A named list of character strings comprised of \code{est} and \code{stat} for each factor.}
   }
}
\description{
These methods take objects from various R functions that calculate ANOVA to create formated chraracter
strings to report the results in accordance with APA manuscript guidelines. \code{anova}-objects from e.g. model comparisons are currently
not supported.
}
\details{
Currently, methods for the following objects exist:
   \itemize{
     \item{\code{aov}}
     \item{\code{summary.aov}}
     \item{\code{aovlist}}
     \item{\code{summary.aovlist}}
     \item{\code{anova}}
   }

   The factor names are sanitized to facilitate their use as list names (see Value section). Parentheses
   are omitted and other non-word characters are replaced by \code{_}.

   If \code{in_paren} is \code{TRUE} parentheses in the formated string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   apa_print(npk_aov)
}
\seealso{
\code{\link{aov}}

Other apa_print: \code{\link{apa_print.htest}};
  \code{\link{apa_print.lm}},
  \code{\link{apa_print.summary.lm}};
  \code{\link{apa_print}}
}

