% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_anova.R, R/apa_print_manova.R
\name{apa_print.aov}
\alias{apa_print.aov}
\alias{apa_print.summary.aov}
\alias{apa_print.aovlist}
\alias{apa_print.summary.aovlist}
\alias{apa_print.Anova.mlm}
\alias{apa_print.summary.Anova.mlm}
\alias{apa_print.afex_aov}
\alias{apa_print.anova}
\alias{apa_print.manova}
\alias{apa_print.summary.manova}
\title{Format statistics from ANOVA (APA 6th edition)}
\usage{
\method{apa_print}{aov}(x, ...)

\method{apa_print}{summary.aov}(x, ...)

\method{apa_print}{aovlist}(x, ...)

\method{apa_print}{summary.aovlist}(
  x,
  estimate = getOption("papaja.estimate_anova", "ges"),
  mse = TRUE,
  observed = NULL,
  intercept = FALSE,
  in_paren = FALSE,
  ...
)

\method{apa_print}{Anova.mlm}(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  ...
)

\method{apa_print}{summary.Anova.mlm}(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  estimate = getOption("papaja.estimate_anova", "ges"),
  mse = getOption("papaja.mse"),
  observed = NULL,
  in_paren = FALSE,
  ...
)

\method{apa_print}{afex_aov}(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  ...
)

\method{apa_print}{anova}(
  x,
  intercept = FALSE,
  estimate = getOption("papaja.estimate_anova", "ges"),
  mse = TRUE,
  observed = NULL,
  in_paren = FALSE,
  ...
)

\method{apa_print}{manova}(x, test = "Pillai", in_paren = FALSE, ...)

\method{apa_print}{summary.manova}(x, in_paren = FALSE, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{...}{[\code{expressions}]\cr Unnamed arguments are taken to be expressions
string(s) to format. Multiple inputs are concatenated together before formatting.
Named arguments are taken to be temporary variables available for substitution.}

\item{estimate}{Character, function, or data frame. Determines which estimate of effect size is to be used. See details.}

\item{mse}{Logical. Indicates if mean squared errors should be included in output. Default is \code{TRUE}.}

\item{observed}{Character. The names of the factors that are observed, (i.e., not manipulated). Necessary for calculation of generalized eta-squared; otherwise ignored.}

\item{intercept}{Logical. Indicates if the intercept term should be included in output.}

\item{in_paren}{Logical. Indicates if the formatted string will be reported inside parentheses. See details.}

\item{correction}{Character. In the case of repeated-measures ANOVA, the type of sphericity correction to be used (\code{"GG"} for Greenhouse-Geisser or \code{"HF"} for Huyn-Feldt methods or \code{"none"}). Default is \code{"GG"}.}

\item{test}{For MANOVA, the multivariate test statistic to be reported, see \code{\link[stats]{summary.manova}}.}
}
\value{
\code{apa_print.aov} and related functions return a named list containing the following components according to the input:

   \describe{
     \item{\code{statistic}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{estimate}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full_result}}{A named list of character strings comprised of \code{estimate} and \code{statistic} for each factor.}
     \item{\code{table}}{A data.frame containing the complete ANOVA table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
These methods take objects from various R functions that calculate ANOVA to create formatted character
strings to report the results in accordance with APA manuscript guidelines. For \code{anova}-objects from model comparisons see \code{\link{apa_print.list}}.
}
\details{
The factor names are sanitized to facilitate their use as list names (see Value section). Parentheses
   are omitted and other non-word characters are replaced by \code{_}.

   If \code{in_paren} is \code{TRUE} parentheses in the formatted string, such as those surrounding degrees
   of freedom, are replaced with brackets.

   Argument `estimate` determines which measure of effect size is to be used: It is currently possible to provide
   a character, where `"ges"` calculates generalized eta squared,`"pes"` calculates partial eta squared, and `"es"`
   calculates eta squared. Note that eta squared is calculated correctly if and only if the design is balanced.

   It is also possible to provide a `data frame` with columns `estimate`, `conf.low`, and `conf.high`, which allows
   for including custom effect-size measures.

   A third option is to provide a function from the \pkg{effectsize} package
   that will be used to calculate effect-size measures from `x`. See the
   examples section for an example.
}
\examples{
   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   apa_print(npk_aov)

   # Use the effectsize package to calculate partial eta-squared with
   # confidence intervals
   apa_print(npk_aov, estimate = effectsize::eta_squared)
}
\references{
Bakeman, R. (2005). Recommended effect size statistics for repeated measures designs. \emph{Behavior Research Methods}
   , 37 (3), 379--384. doi:\href{https://doi.org/10.3758/BF03192707}{10.3758/BF03192707}
}
\seealso{
\code{\link{aov}}, \code{\link[car]{Anova}}, \code{\link{apa_print.list}}

Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.emmGrid}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print.lme}()},
\code{\link{apa_print.merMod}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
