% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_anova.R
\name{apa_print.aov}
\alias{apa_print.aov}
\alias{apa_print.summary.aov}
\alias{apa_print.aovlist}
\alias{apa_print.summary.aovlist}
\alias{apa_print.Anova.mlm}
\alias{apa_print.summary.Anova.mlm}
\alias{apa_print.afex_aov}
\alias{apa_print.anova}
\title{Format statistics from ANOVA (APA 6th edition)}
\usage{
apa_print.aov(x, ...)

apa_print.summary.aov(x, ...)

apa_print.aovlist(x, ...)

apa_print.summary.aovlist(x, ...)

apa_print.Anova.mlm(x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE, ...)

apa_print.summary.Anova.mlm(x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE, ...)

apa_print.afex_aov(x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE, ...)

apa_print.anova(x, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{...}{Arguments passed on to \code{print_anova}
\describe{
  \item{x}{Data.frame. A \code{data.frame} of class \code{apa_variance_table} as returned by \code{\link{arrange_anova}}.}
  \item{intercept}{Logical. Indicates if intercept test should be included in output.}
  \item{es}{Character. The effect-size measure to be calculated; can be either \code{ges} for generalized eta-squared, \code{pes} for partial eta-squared or \code{es} for eta-squared.
Note that eta-squared is calculated correctly if and only if the design is balanced.}
  \item{mse}{Logical. Indicates if mean squared errors should be included in output. Default is \code{TRUE}.}
  \item{observed}{Character. The names of the factors that are observed, (i.e., not manipulated). Necessary for calculation of generalized eta-squared; otherwise ignored.}
  \item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}
}}

\item{correction}{Character. In the case of repeated-measures ANOVA, the type of sphericity correction to be used (\code{GG} for Greenhouse-Geisser or \code{HF} for Huyn-Feldt methods or \code{none}). Default is \code{GG}.}

\item{intercept}{Logical. Indicates if intercept test should be included in output.}
}
\value{
\code{apa_print.aov} and related functions return a named list containing the following components according to the input:

   \describe{
     \item{\code{statistic}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{estimate}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full_result}}{A named list of character strings comprised of \code{estimate} and \code{statistic} for each factor.}
     \item{\code{table}}{A data.frame containing the complete ANOVA table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
These methods take objects from various R functions that calculate ANOVA to create formatted character
strings to report the results in accordance with APA manuscript guidelines. For \code{anova}-objects from model comparisons see \code{\link{apa_print.list}}.
}
\details{
The factor names are sanitized to facilitate their use as list names (see Value section). Parentheses
   are omitted and other non-word characters are replaced by \code{_}.

   If \code{in_paren} is \code{TRUE} parentheses in the formatted string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   apa_print(npk_aov)
}
\references{
Bakeman, R. (2005). Recommended effect size statistics for repeated measures designs. \emph{Behavior Research Methods}
   , 37 (3), 379--384. doi:\href{https://doi.org/10.3758/BF03192707}{10.3758/BF03192707}
}
\seealso{
\code{\link{aov}}, \code{\link[car]{Anova}}, \code{\link{apa_print.list}}

Other apa_print: \code{\link{apa_print.BFBayesFactor}},
  \code{\link{apa_print.emmGrid}},
  \code{\link{apa_print.glht}},
  \code{\link{apa_print.glm}},
  \code{\link{apa_print.htest}},
  \code{\link{apa_print.list}}, \code{\link{apa_print}}
}
\concept{apa_print}
