% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_anova.R
\name{apa_print.aov}
\alias{apa_print.Anova.mlm}
\alias{apa_print.afex_aov}
\alias{apa_print.anova}
\alias{apa_print.aov}
\alias{apa_print.aovlist}
\alias{apa_print.summary.Anova.mlm}
\alias{apa_print.summary.aov}
\alias{apa_print.summary.aovlist}
\title{Format statistics from ANOVA (APA 6th edition)}
\usage{
\method{apa_print}{aov}(x, ...)

\method{apa_print}{summary.aov}(x, ...)

\method{apa_print}{aovlist}(x, ...)

\method{apa_print}{summary.aovlist}(x, ...)

\method{apa_print}{Anova.mlm}(x, correction = "GG", intercept = FALSE, ...)

\method{apa_print}{summary.Anova.mlm}(x, correction = "GG",
  intercept = FALSE, ...)

\method{apa_print}{afex_aov}(x, correction = "GG", intercept = FALSE, ...)

\method{apa_print}{anova}(x, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{...}{Additional arguments passed to or from other methods.}

\item{correction}{Character. In the case of repeated-measures ANOVA, the type of sphericity correction to be used. Either \code{GG} for Greenhouse-Geisser or \code{HF} for Huyn-Feldt methods or \code{none} is also possible.}

\item{intercept}{Logical. Indicates if intercept test should be included in output.}

\item{es}{Character. The effect-size measure to be calculated; can be either \code{ges} for generalized eta-squared or \code{pes} for partial eta-squared.}

\item{observed}{Character. The names of the factors that are observed, (i.e., not manipulated). Necessary for calculation of generalized eta-squared; otherwise ignored.}

\item{mse}{Logical. Specifies if Mean Squared Errors (MSEs) are to be reported.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}
}
\value{
\code{apa_print.aov} and related functions return a named list containing the following components according to the input:

   \describe{
     \item{\code{stat}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{est}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A named list of character strings comprised of \code{est} and \code{stat} for each factor.}
     \item{\code{table}}{A data.frame containing the complete ANOVA table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
These methods take objects from various R functions that calculate ANOVA to create formatted chraracter
strings to report the results in accordance with APA manuscript guidelines. For \code{anova}-objects from model comparisons see \code{\link{apa_print.list}}.
}
\details{
The factor names are sanitized to facilitate their use as list names (see Value section). Parentheses
   are omitted and other non-word characters are replaced by \code{_}.

   If \code{in_paren} is \code{TRUE} parentheses in the formatted string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   apa_print(npk_aov)
}
\references{
Bakeman, R. (2005). Recommended effect size statistics for repeated measures designs. \emph{Behavior Research Methods}
   , 37 (3), 379--384. doi:\href{http://dx.doi.org/10.3758/BF03192707}{10.3758/BF03192707}
}
\seealso{
\code{\link{aov}}, \code{\link[car]{Anova}}, \code{\link{apa_print.list}}

Other apa_print: \code{\link{apa_print.glht}},
  \code{\link{apa_print.htest}},
  \code{\link{apa_print.list}}, \code{\link{apa_print.lm}},
  \code{\link{apa_print}}
}

