% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_BFBayesFactor.R
\docType{methods}
\name{apa_print.BFBayesFactor}
\alias{apa_print,BFBayesFactor-method}
\alias{apa_print,BFBayesFactorList-method}
\alias{apa_print.BFBayesFactor}
\alias{apa_print.BFBayesFactorList}
\title{Format Bayes factors (APA 6th edition)}
\usage{
apa_print.BFBayesFactor(x, ratio_subscript = "10", auto_invert = TRUE,
  scientific = TRUE, ...)

\S4method{apa_print}{BFBayesFactor}(x, ratio_subscript = "10",
  auto_invert = TRUE, scientific = TRUE, ...)

apa_print.BFBayesFactorList(x, ...)

\S4method{apa_print}{BFBayesFactorList}(x, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{ratio_subscript}{Character. A brief description of the model comparison in the form of \code{"M1/M2"}.}

\item{scientific}{Logical. Indicates whether to use scientific notation.}

\item{...}{}

\item{auto_inverse}{Logical. Indicates whether the Bayes factor should be inverted (including \code{ratio_subscript}) if it is less than 1.}
}
\value{
...
}
\description{
These methods take result objects from the \code{\pkg{BayesFactor}} package to create
 formatted chraracter strings to report the results in accordance with APA manuscript
 guidelines.
}
\examples{
data(sleep)
bayesian_anova <- anovaBF(
  extra ~ group + ID
  , data = sleep
  , whichRandom = "ID"
  , progress=FALSE
)

apa_print(bayesian_anova)
}
\seealso{
Other apa_print: \code{\link{apa_print.aov}},
  \code{\link{apa_print.glht}},
  \code{\link{apa_print.htest}},
  \code{\link{apa_print.list}}, \code{\link{apa_print.lm}},
  \code{\link{apa_print}}
}

