% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_BFBayesFactor.R
\name{apa_print.BFBayesFactor}
\alias{apa_print.BFBayesFactor}
\title{Title}
\usage{
\method{apa_print}{BFBayesFactor}(x, ratio_subscript = "10",
  auto_invert = TRUE, scientific = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

