% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_emm_lsm.R
\name{apa_print.emmGrid}
\alias{apa_print.emmGrid}
\alias{apa_print.summary_emm}
\alias{apa_print.lsmobj}
\alias{apa_print.summary.ref.grid}
\title{Typeset Statistical Results from Estimated Marginal Means}
\usage{
\method{apa_print}{emmGrid}(x, ...)

\method{apa_print}{summary_emm}(x, contrast_names = NULL, est_name = NULL, in_paren = FALSE, ...)

\method{apa_print}{lsmobj}(x, ...)

\method{apa_print}{summary.ref.grid}(x, ...)
}
\arguments{
\item{x}{Object
adaptation, such as \eqn{\chi^2} instead of "x-squared") is used for the \emph{test statistic}, otherwise the}

\item{...}{
  Arguments passed on to \code{\link[=printnum]{printnum}}
  \describe{
    \item{\code{}}{}
  }}

\item{contrast_names}{Character. An optional vector of names to identify calculated contrasts.}

\item{est_name}{Character. If \code{NULL} (default) the name is guessed from the function call of the model object passed to \code{lsmeans}/\code{emmeans}.}

\item{in_paren}{Logical. Whether the formatted string is to be reported in
parentheses. If \code{TRUE}, parentheses in the formatted string (e.g., those
enclosing degrees of freedom) are replaced with brackets.}
}
\description{
Takes various \code{lsmeans} and \code{emmeans} objects to create formatted character strings to report the results in
accordance with APA manuscript guidelines.  \emph{These methods are not properly tested and should be
considered experimental.}
}
\details{
ADJUSTED CONFIDENCE INTERVALS
}
\examples{
   NULL
}
\seealso{
Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.aov}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print.lme}()},
\code{\link{apa_print.merMod}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
\value{
\code{apa_print()}-methods return a named list of class \code{apa_results} containing the following elements:

\item{estimate}{One or more character strings giving point estimates, confidence intervals, and confidence level. A single string is returned in a vector; multiple strings are returned as a named list. If no estimate is available the element is \code{NULL}.}
\item{statistic}{One or more character strings giving the test statistic, parameters (e.g., degrees of freedom), and p-value. A single string is returned in a vector; multiple strings are returned as a named list. If no estimate is available the element is \code{NULL}.}
\item{full_result}{One or more character strings comprised `estimate` and `statistic`. A single string is returned in a vector; multiple strings are returned as a named list. }
\item{table}{A \code{data.frame} of class \code{apa_results_table} that contains all elements of \code{estimate} and \code{statistics}. This table can be passed to \code{apa_table()} for reporting.}

Column names in \code{apa_results_table} are standardized following the \href{https://www.tidymodels.org/learn/develop/broom/#glossary}{\pkg{broom} glossary} (e.g., \code{term}, \code{estimate} \code{conf.int}, \code{statistic}, \code{df}, \code{df.residual}, \code{p.value}). Additionally, each column is labelled (e.g., \code{$\\hat{\\eta}^2_G$} or \code{$t(23)$}) using the \pkg{tinylabels} package and these labels are used as column names when an \code{apa_results_table} is passed to \code{apa_table()}.

}
