% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_emm_lsm.R
\name{apa_print.emmGrid}
\alias{apa_print.emmGrid}
\alias{apa_print.summary_emm}
\alias{apa_print.lsmobj}
\alias{apa_print.summary.ref.grid}
\title{Format statistics (APA 6th edition)}
\usage{
\method{apa_print}{emmGrid}(x, infer = TRUE, conf.level = 0.95, ...)

\method{apa_print}{summary_emm}(
  x,
  contrast_names = NULL,
  est_name = "\\\\hat{\\\\theta}",
  in_paren = FALSE,
  ...
)

\method{apa_print}{lsmobj}(x, ...)

\method{apa_print}{summary.ref.grid}(x, ...)
}
\arguments{
\item{x}{Object}

\item{infer}{A vector of one or two logical values. The first determines
whether confidence intervals are displayed, and the second determines
whether \emph{t} tests and \emph{P} values are displayed. If only one value
is provided, it is used for both.}

\item{conf.level}{Numeric. Confidence level for confidence intervals.}

\item{...}{
  Arguments passed on to \code{\link[=printnum]{printnum}}
  \describe{
    \item{\code{}}{}
  }}

\item{contrast_names}{Character. An optional vector of names to label the
calculated contrasts.}

\item{est_name}{Character. If \code{NULL} (default) the name of the estimate
is inferred from the function call of the model object supplied to
\pkg{emmeans}.}

\item{in_paren}{Logical. Whether the formatted string is to be reported in
parentheses. If \code{TRUE}, parentheses in the formatted string (e.g., those
enclosing degrees of freedom) are replaced with brackets.}
}
\value{
\code{apa_print()} returns a list containing the following components according to the input:

\describe{
\item{\code{statistic}}{A character string giving the test statistic, parameters (e.g., degrees of freedom),
and \emph{p} value.}
\item{\code{estimate}}{A character string giving the descriptive estimates and confidence intervals if possible}
\% , either in units of the analyzed scale or as standardized effect size.
\item{\code{full_result}}{A joint character string comprised of \code{est} and \code{stat}.}
\item{\code{table}}{A data.frame containing the complete contrast table, which can be passed to \code{\link{apa_table}}.}
}
}
\description{
Takes various \pkg{emmeans} objects to create formatted character strings to
report the results in accordance with APA manuscript guidelines.
\emph{\pkg{emmeans} supports a wide range of analysis, not all
of which are currently (fully) supported. Proceed with caution.}
}
\details{
When p-values and confidence intervals are adjusted for multiple testing,
the correction method is added as an index to the output (e.g.
\verb{p_\{Tukey(3)\}}). Values in parenthesis indicate the size of the family of
tests or the rank of the set of linear functions (for the Scheffé method).

If possible, each family of tests is additionally marked in the returned
table by alphabetic superscripts.

Generally, the \code{summary_emm} objects returned by \code{emmeans::summary_emm} omit
information that may be needed to add some of the information on the
adjustments made to p-values and confidence intervals. It is therefore
preferable to pass \code{emmGrid}-objects if possible. For example, by using
\code{emmeans(object, 1 ~ x1, adjust = "scheffe")}.
}
\examples{
   NULL
}
\seealso{
Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.aov}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print.lme}()},
\code{\link{apa_print.merMod}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
