% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_emm_lsm.R
\name{apa_print.emmGrid}
\alias{apa_print.emmGrid}
\alias{apa_print.summary_emm}
\alias{apa_print.lsmobj}
\alias{apa_print.summary.ref.grid}
\title{Format statistics (APA 6th edition)}
\usage{
apa_print.emmGrid(x, ...)

apa_print.summary_emm(x, contrast_names = NULL, est_name = NULL,
  in_paren = FALSE, ...)

apa_print.lsmobj(x, ...)

apa_print.summary.ref.grid(x, ...)
}
\arguments{
\item{x}{Object
adaptation, such as \eqn{\chi^2} instead of "x-squared") is used for the \emph{test statistic}, otherwise the}

\item{...}{Arguments passed on to \code{printnum}
\describe{
  \item{x}{Numeric. Can be either a single value, vector, or matrix.}
}}

\item{contrast_names}{Character. An optional vector of names to identify calculated contrasts.}

\item{est_name}{Character. If \code{NULL} (default) the name is guessed from the function call of the model object passed to \code{lsmeans}/\code{emmeans}.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses.}
}
\value{
\code{apa_print()} returns a list containing the following components according to the input:

   \describe{
     \item{\code{statistic}}{A character string giving the test statistic, parameters (e.g., degrees of freedom),
         and \emph{p} value.}
     \item{\code{estimate}}{A character string giving the descriptive estimates and confidence intervals if possible}
         % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full_result}}{A joint character string comprised of \code{est} and \code{stat}.}
     \item{\code{table}}{A data.frame containing the complete contrast table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
Takes various \code{lsmeans} and \code{emmeans} objects to create formatted character strings to report the results in
accordance with APA manuscript guidelines.  \emph{These methods are not properly tested and should be
considered experimental.}
}
\details{
ADJUSTED CONFIDENCE INTERVALS

   If \code{in_paren} is \code{TRUE} parentheses in the formated string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
   NULL
}
\seealso{
Other apa_print: \code{\link{apa_print.BFBayesFactor}},
  \code{\link{apa_print.aov}},
  \code{\link{apa_print.glht}},
  \code{\link{apa_print.glm}},
  \code{\link{apa_print.htest}},
  \code{\link{apa_print.list}}, \code{\link{apa_print}}
}
\concept{apa_print}
