% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_print_glht.R
\name{apa_print.glht}
\alias{apa_print.glht}
\alias{apa_print.lsmobj}
\alias{apa_print.summary.glht}
\alias{apa_print.summary.ref.grid}
\title{Format statistics (APA 6th edition)}
\usage{
\method{apa_print}{glht}(x, test = adjusted(), ...)

\method{apa_print}{summary.glht}(x, ci = 0.95, in_paren = FALSE, ...)

\method{apa_print}{lsmobj}(x, ...)

\method{apa_print}{summary.ref.grid}(x, ci = 0.95, in_paren = FALSE,
  contrast_names = NULL, ...)
}
\arguments{
\item{x}{See details.}

\item{test}{Function.}

\item{...}{Further arguments to pass to \code{\link{printnum}} to format the estimate.}

\item{ci}{Numeric. If \code{NULL} (default) the function tries to obtain confidence intervals from \code{x}.
Other confidence intervals can be supplied as a \code{vector} of length 2 (lower and upper boundary, respectively)
with attribute \code{conf.level}, e.g., when calculating bootstrapped confidence intervals.}
}
\value{
\code{apa_print()} returns a list containing the following components according to the input:

   \describe{
     \item{\code{stat}}{A character string giving the test statistic, parameters (e.g., degrees of freedom),
         and \emph{p} value.}
     \item{\code{est}}{A character string giving the descriptive estimates and confidence intervals if possible}
         % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A joint character string comprised of \code{est} and \code{stat}.}
   }
}
\description{
Takes \code{htest} objects from various statistical methods to create
formated chraracter strings to report the results in accordance with APA manuscript guidelines.
}
\details{
The function should work on a wide range of \code{htest} objects. Due to the large number of functions
   that produce these objects and their idiosyncracies, the produced strings may sometimes be inaccurate. If you
   experience inaccuracies you may report these \href{https://github.com/crsh/papaja/issues}{here} (please include
   a reproducible example in your report!).

   ADJUSTED CONFIDENCE INTERVALS

   \code{stat_name} and \code{est_name} are placed in the output string and are thus passed to pandoc or LaTeX through
   \pkg{kntir}. Thus, to the extent it is supported by the final document type, you can pass LaTeX-markup to format the
   final text (e.g., \code{\\\\tau} yields \eqn{\tau}).

   If \code{in_paren} is \code{TRUE} parentheses in the formated string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
NULL
}
\seealso{
Other apa_print: \code{\link{apa_print.Anova.mlm}},
  \code{\link{apa_print.afex_aov}},
  \code{\link{apa_print.anova}},
  \code{\link{apa_print.aov}},
  \code{\link{apa_print.aovlist}},
  \code{\link{apa_print.summary.Anova.mlm}},
  \code{\link{apa_print.summary.aov}},
  \code{\link{apa_print.summary.aovlist}};
  \code{\link{apa_print.htest}};
  \code{\link{apa_print.lm}},
  \code{\link{apa_print.summary.lm}};
  \code{\link{apa_print}}
}

