% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apa_print_htest.R
\name{apa_print.htest}
\alias{apa_print.htest}
\title{Format statistics (APA 6th edition)}
\usage{
\method{apa_print}{htest}(x, stat_name = NULL, est_name = NULL, n = NULL,
  ci = NULL, in_paren = FALSE, ...)
}
\arguments{
\item{x}{\code{htest} object. See details.}

\item{stat_name}{Character. If \code{NULL} (default) the name given in \code{x} (or a formally correct
adaptation, such as \eqn{\chi^2} instead of "x-squared") is used for the \emph{test statistic}, otherwise the
supplied name is used. See details.}

\item{est_name}{Character. If \code{NULL} (default) the name given in \code{x} (or a formally correct
adaptation, such as \eqn{r_S} instead of "rho") is used for the \emph{estimate}, otherwise the supplied name is
used. See details.}

\item{n}{Numeric. Size of the sample; required when reporting \eqn{\chi^2} tests, otherwise this parameter
is ignored.}

\item{ci}{Numeric. If \code{NULL} (default) the function tries to obtain confidence intervals from \code{x}.
Other confidence intervals can be supplied as a \code{vector} of length 2 (lower and upper boundary, respectively)
with attribute \code{conf.level}, e.g., when calculating bootstrapped confidence intervals.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}

\item{...}{Further arguments to pass to \code{\link{printnum}} to format the estimate.}
}
\value{
\code{apa_print()} returns a list containing the following components according to the input:

   \describe{
     \item{\code{stat}}{A character string giving the test statistic, parameters (e.g., degrees of freedom),
         and \emph{p} value.}
     \item{\code{est}}{A character string giving the descriptive estimates and confidence intervals if possible}
         % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A joint character string comprised of \code{est} and \code{stat}.}
   }
}
\description{
Takes \code{htest} objects from various statistical methods to create
formated chraracter strings to report the results in accordance with APA manuscript guidelines.
}
\details{
The function should work on a wide range of \code{htest} objects. Due to the large number of functions
   that produce these objects and their idiosyncracies, the produced strings may sometimes be inaccurate. If you
   experience inaccuracies you may report these \href{https://github.com/crsh/papaja/issues}{here} (please include
   a reproducible example in your report!).

   \code{stat_name} and \code{est_name} are placed in the output string and are thus passed to pandoc or LaTeX through
   \pkg{kntir}. Thus, to the extent it is supported by the final document type, you can pass LaTeX-markup to format the
   final text (e.g., \code{\\\\tau} yields \eqn{\tau}).

   If \code{in_paren} is \code{TRUE} parentheses in the formated string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
# Comparisions of central tendencies
t_stat <- t.test(extra ~ group, data = sleep)
apa_print(t_stat)
apa_print(t_stat, stat_name = "tee")

wilcox_stat <- wilcox.test(extra ~ group, data = sleep)
apa_print(wilcox_stat)

# Correlations
## Data from Hollander & Wolfe (1973), p. 187f.
x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
y <- c( 2.6,  3.1,  2.5,  5.0,  3.6,  4.0,  5.2,  2.8,  3.8)
cor_stat <- cor.test(x, y, method = "spearman")
apa_print(cor_stat)

# Contingency tables
## Data from Fleiss (1981), p. 139.
smokers  <- c(83, 90, 129, 70)
patients <- c(86, 93, 136, 82)
prop_stat <- prop.test(smokers, patients)
apa_print(prop_stat, n = sum(patients))
}
\seealso{
Other apa_print: \code{\link{apa_print.Anova.mlm}},
  \code{\link{apa_print.afex_aov}},
  \code{\link{apa_print.anova}},
  \code{\link{apa_print.aov}},
  \code{\link{apa_print.aovlist}},
  \code{\link{apa_print.summary.Anova.mlm}},
  \code{\link{apa_print.summary.aov}},
  \code{\link{apa_print.summary.aovlist}};
  \code{\link{apa_print.lm}},
  \code{\link{apa_print.summary.lm}};
  \code{\link{apa_print}}
}

