% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_list.R
\name{apa_print.list}
\alias{apa_print.list}
\title{Format statistics from ANOVA (APA 6th edition)}
\usage{
\method{apa_print}{list}(x, anova_fun = anova, ci = 0.9,
  boot_samples = 10000, in_paren = FALSE, ...)
}
\arguments{
\item{x}{List. List containing to be compared \code{lm}-objects. If the list is completely named, element names are used as model names in the ouptut object.}

\item{anova_fun}{Function. Function to compare model-objects contained in \code{x}.}

\item{ci}{Numeric. Confidence level for the bootstrap confidence interval for \eqn{\Delta R^2} (range [0, 1]); ignored if \code{boot_samples = 0}.}

\item{boot_samples}{Numeric. Number of bootstrap samples to estimate confidence intervals for \eqn{\Delta R^2}.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}

\item{...}{}
}
\value{
\code{apa_print.list} returns a named list containing the following components according to the input:

   \describe{
     \item{\code{stat}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each non-baseline model.}
     \item{\code{est}}{A named list of character strings giving the effect size estimates for each non-baseline model}
     \item{\code{full}}{A named list of character strings comprised of \code{est} and \code{stat} for each non-baseline model}
     \item{\code{table}}{A data.frame containing the complete model comparison table including regression coefficients, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
This methods performs comparisons of \code{lm}-objects and creates formatted chraracter
strings and a model comparison table to report the results in accordance with APA manuscript guidelines.
}
\details{
As demonstrated by Algina, Keselman & Penfield (2007), asymptotic confidence intervals for \eqn{\Delta R^2}
   are often unreliable. Confidence intervals for model comparisons of \code{lm}-objects are, therefore, estimated
   using their modified percentile bootstrap method. Note that the accuracy of the confidence intervals depends on
   the number of predictors \eqn{p}, their distribution, and the sample size \eqn{n}:

   \emph{"When the predictor distribution is multivariate normal, one can obtain accurate CIs for \eqn{\rho^2} with
   \eqn{n \geq~50} when \eqn{p = 3}. For \eqn{p = 6} and for \eqn{p = 9}, \eqn{n \geq~100} is advisable. When the
   predictor distribution is nonnormal in form, sample size requirements vary with type of nonnormality." (p. 939,
   Algina, Keselman & Penfield, 2010)}

   Confidence intervals for \eqn{R^2} and \eqn{\Delta R^2} are computed for the requested confidence level (\code{ci}), default being a 90\% CI, see Steiger (2004).
   The confidence intervals for the regression coefficients in the model comparison table are correspond to the \eqn{\alpha}-level chosen for \eqn{R^2} and \eqn{\Delta R^2} (e.g., 90\% CI or \eqn{\alpha = 0.10} for \eqn{R^2} and \eqn{\Delta R^2} yields a 95\% CI for regression coefficients, Steiger, 2004)

   If \code{in_paren} is \code{TRUE} parentheses in the formatted string, such as those surrounding degrees
   of freedom, are replaced with brackets.
}
\examples{
   mod1 <- lm(Sepal.Length ~ Sepal.Width, data = iris)
   mod2 <- update(mod1, formula = . ~ . + Petal.Length)
   mod3 <- update(mod2, formula = . ~ . + Petal.Width)

   # No bootstrapped Delta R^2 CI
   apa_print(list(Baseline = mod1, Length = mod2, Both = mod3), boot_samples = 0)
}
\references{
Algina, J., Keselman, H. J., & Penfield, R. D. (2007). Confidence Intervals for an Effect Size Measure in Multiple Linear Regression.
   \emph{Educational and Psychological Measurement}, 67(2), 207--218. doi:\href{http://dx.doi.org/10.1177/0013164406292030}{10.1177/0013164406292030}

   Algina, J., Keselman, H. J., & Penfield, R. D. (2010). Confidence Intervals for Squared Semipartial Correlation Coefficients: The Effect of Nonnormality.
   \emph{Educational and Psychological Measurement}, 70(6), 926--940. doi:\href{http://dx.doi.org/10.1177/0013164410379335}{10.1177/0013164410379335}

Steiger (2004). Beyond the F Test: Effect Size Confidence Intervals and Tests of Close Fit in the Analysis of
Variance and Contrast Analysis. \emph{Psychological Methods}, 9(2), 164-182.
doi:\href{http://dx.doi.org/10.1037/1082-989X.9.2.164}{10.1037/1082-989X.9.2.164}
}
\seealso{
\code{\link{anova}}

Other apa_print: \code{\link{apa_print.aov}},
  \code{\link{apa_print.glht}},
  \code{\link{apa_print.htest}},
  \code{\link{apa_print.lm}}, \code{\link{apa_print}}
}

