% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_nlme.R
\name{apa_print.lme}
\alias{apa_print.lme}
\alias{apa_print.anova.lme}
\title{Format Statistics from Nonlinear Mixed-Effects Models (APA 6th edition)}
\usage{
\method{apa_print}{lme}(x, args_confint = NULL, in_paren = FALSE, est_name = NULL, ...)

\method{apa_print}{anova.lme}(x, in_paren = FALSE, ...)
}
\arguments{
\item{x}{A (non-)linear mixed-effects model fitted with \code{\link[nlme:lme]{nlme::lme()}} or \code{\link[nlme:nlme]{nlme::nlme()}}}

\item{args_confint}{Named list. Arguments that are passed to \code{\link[nlme:intervals.lme]{nlme::intervals.lme()}}.}

\item{in_paren}{Logical. Whether the formatted string is to be reported in
parentheses. If `TRUE`, parentheses in the formatted string (e.g., those
enclosing degrees of freedom) are replaced with brackets.}

\item{est_name}{An optional character. The label to be used for fixed-effects coefficients.}

\item{...}{Further arguments that are passed to \code{\link[=printnum]{printnum()}}.}
}
\description{
These methods take mixed-effects models fitted with the \pkg{nlme} package
and create formatted character strings report the results in accordance
with APA manuscript guidelines.
}
\examples{
  library(nlme)
  fm1 <- lme(distance ~ age, data = Orthodont, method = "ML") # random is ~ age
  apa_print(fm1)
  # ANOVA-like tables
  single_anova <- anova(fm1)
  apa_print(single_anova)

}
\seealso{
Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.aov}()},
\code{\link{apa_print.emmGrid}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print.merMod}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
