% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_merMod.R
\name{apa_print.merMod}
\alias{apa_print.merMod}
\alias{apa_print.mixed}
\title{Format Statistics from (Generalized) Hierarchical Linear Models (APA 6th edition)}
\usage{
\method{apa_print}{merMod}(x, args_confint, ...)

\method{apa_print}{mixed}(x, ...)
}
\arguments{
\item{x}{A fitted hierarchical linear model, either from \code{\link[lme4:lmer]{lme4::lmer()}},
\code{\link[lmerTest:lmer]{lmerTest::lmer()}}, or \code{\link[afex:mixed]{afex::mixed()}}.}

\item{args_confint}{Named list. Additional arguments that are passed to \code{\link[lme4:confint.merMod]{lme4::confint.merMod()}}.}

\item{...}{Further arguments, currently ignored.}
}
\description{
These methods take objects from various R functions that calculate
hierarchical linear models to create formatted character strings
to report the results in accordance with APA manuscript guidelines.
}
\seealso{
Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.aov}()},
\code{\link{apa_print.emmGrid}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
