% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_merMod.R
\name{apa_print.merMod}
\alias{apa_print.merMod}
\alias{apa_print.mixed}
\title{Format Statistics from Hierarchical (Generalized) Linear Models (APA 6th edition)}
\usage{
\method{apa_print}{merMod}(
  x,
  effects = "fixed",
  args_confint = NULL,
  in_paren = FALSE,
  est_name = NULL,
  ...
)

\method{apa_print}{mixed}(x, ...)
}
\arguments{
\item{x}{A fitted hierarchical (generalized) linear model, either from \code{\link[lme4:lmer]{lme4::lmer()}},
\code{\link[lmerTest:lmer]{lmerTest::lmer()}}, \code{\link[afex:mixed]{afex::mixed()}}, or \code{\link[lme4:glmer]{lme4::glmer()}}.}

\item{effects}{Character. Determines which information is returned. Currently, only fixed-effects terms
(\code{"fixed"}) are supported.}

\item{args_confint}{Named list. Additional arguments that are passed to \code{\link[lme4:confint.merMod]{lme4::confint.merMod()}}, see details.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}

\item{est_name}{An optional character. The label to be used for fixed-effects coefficients.}

\item{...}{Further arguments that are passed to \code{\link[=printnum]{printnum()}}.}
}
\description{
These methods take objects from various R functions that calculate
hierarchical (generalized) linear models to create formatted character strings
to report the results in accordance with APA manuscript guidelines.
}
\details{
Confidence intervals are calculated by calling \code{\link[lme4:confint.merMod]{lme4::confint.merMod()}}. By
default, \emph{Wald} confidence intervals are calculated, but this may change in
the future.
}
\examples{
  # Fit a linear mixed model using the lme4 package
  # or the lmerTest package (for dfs and p values)
  library(lmerTest)
  fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
  # Format statistics for fixed-effects terms (the default)
  apa_print(fm1)

}
\seealso{
Other apa_print: 
\code{\link{apa_print.BFBayesFactor}()},
\code{\link{apa_print.aov}()},
\code{\link{apa_print.emmGrid}()},
\code{\link{apa_print.glht}()},
\code{\link{apa_print.glm}()},
\code{\link{apa_print.htest}()},
\code{\link{apa_print.list}()},
\code{\link{apa_print.lme}()},
\code{\link{apa_print}()}
}
\concept{apa_print}
