% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_wsci.R
\name{apa_print.papaja_wsci}
\alias{apa_print.papaja_wsci}
\title{Format Within-Subjects Confidence Intervals (APA 6th edition)}
\usage{
\method{apa_print}{papaja_wsci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{papaja_wsci}.}

\item{...}{Arguments passed on to \code{\link{printnum}}}
}
\description{
This method takes an output object from \code{\link{wsci}} and creates a table
and character strings to report means and within-subjects confidence intervals
in a table or in text.
}
