% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print.R
\name{apa_print}
\alias{apa_print}
\title{Format statistics (APA 6th edition)}
\usage{
apa_print(x, ...)
}
\arguments{
\item{x}{Output object. See details.}

\item{...}{Additional arguments passed to other methods. See details.}
}
\value{
The form of the value returned by \code{apa_print} depends on the class of \code{x}. See the
   documentation of the particular methods for details (e.g., \code{?apa_print.htest}).
}
\description{
A generic function that takes objects from various statistical methods to create formatted chraracter
strings to report the results in accordance with APA manuscript guidelines. The function invokes particular
methods, which depend on the \code{\link{class}} of the first argument.
}
\details{
Currently the following output objects are supported:

   \itemize{
     \item \code{htest}
     \item \code{lm} and \code{summary.lm}
     \item \code{aov}, \code{aovlist}, \code{summary.aov}, and \code{summary.aovlist}
     \item \code{anova} and \code{Anova.mlm}
   }
}
\examples{
# List methods for apa_print()
methods("apa_print")
}
\seealso{
Other apa_print: \code{\link{apa_print.aov}},
  \code{\link{apa_print.glht}},
  \code{\link{apa_print.htest}}, \code{\link{apa_print.lm}}
}

