% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{apa_table}
\alias{apa_table}
\alias{apa_table.latex}
\alias{apa_table.word}
\title{Prepare table for printing}
\usage{
apa_table(x, caption = NULL, note = NULL, stub_indents = NULL,
  added_stub_head = NULL, col_spanners = NULL, midrules = NULL,
  placement = "tbp", landscape = FALSE, small = FALSE, ...)

apa_table.latex(x, caption = NULL, note = NULL, col_spanners = NULL,
  midrules = NULL, placement = "tbp", landscape = FALSE, small = FALSE,
  ...)

apa_table.word(x, caption = NULL, note = NULL, ...)
}
\arguments{
\item{x}{Object to print, can be \code{matrix}, \code{data.frame}, or \code{list}. See details.}

\item{caption}{Character. Caption to be printed above the table.}

\item{note}{Character. Note to be printed below the table.}

\item{stub_indents}{List. A named list of vectors of length 2 giving the first and second row to
indent. Names of list elements will be used as titles for indented sections.}

\item{added_stub_head}{Character. Used as stub head (name of first column) if \code{row.names = TRUE}
is passed to \code{\link[knitr]{kable}}; ignored if row names are omitted from the table.}

\item{col_spanners}{List. A named list of vectors of length 2 giving the first and second column to
span with a grouping column name.}

\item{midrules}{Numeric. Vector of line numbers in table (not counting column headings) that should be
followed by a horizontal rule; ignored in MS Word documents.}

\item{placement}{Character. Indicates whether table should be placed at the exact location (\code{h}),
at the top (\code{t}), bottom (\code{b}), or on a separate page (\code{p}). Arguments can be combined
to indicate order of preference (\code{htb}); ignored in MS Word documents.}

\item{landscape}{Logical. If \code{TRUE} the table is printed in landscape format; ignored in MS Word
documents.}

\item{small}{Logical. If \code{TRUE} the font size of the table content is reduced.}

\item{...}{Further arguments to pass to \code{\link[knitr]{kable}}.}
}
\description{
Formats \code{matrices} and \code{data.frames} to report them as tables according to APA guidelines
(6th edition).
}
\details{
When using \code{apa_table()}, the type of the ouput (LaTeX or MS Word) is determined automatically
   by the rendered document type. If no rendering is in progress the output default is LaTeX.

   If \code{x} is a \code{list}, all list elements are merged by columns into a single table with
   the first column giving the names of the list elements elements.
}
\examples{

my_table <- apply(cars, 2, function(x) # Create data
   round(c(Mean = mean(x), SD = sd(x), Min = min(x), Max = max(x)), 2)
)

apa_table(
   my_table
   , align = c("l", "r", "r")
   , caption = "A summary table of the cars dataset."
   , note = "This table was created using apa\\\\_table()"
   , added_stub_head = "Descriptives"
)
}
\seealso{
\code{\link[knitr]{kable}}
}

