% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{beautify_terms}
\alias{beautify_terms}
\alias{beautify_terms.character}
\alias{beautify_terms.numeric}
\alias{beautify_terms.factor}
\alias{beautify_terms.data.frame}
\title{Prettify Term Names}
\usage{
beautify_terms(x, ...)

\method{beautify_terms}{character}(x, standardized = FALSE, retain_period = FALSE, ...)

\method{beautify_terms}{numeric}(x, standardized = FALSE, ...)

\method{beautify_terms}{factor}(x, standardized = FALSE, ...)

\method{beautify_terms}{data.frame}(x, ...)
}
\arguments{
\item{x}{Character. Vector of term names to be prettified.}

\item{standardized}{Logical. If \code{TRUE}, the name of the function \code{\link[=scale]{scale()}} will be
removed from term names.}
}
\description{
Remove parentheses, replace colons with \code{$\\times$}.
Useful to prettify term names in \code{\link[=apa_print]{apa_print()}} tables.
}
\examples{
NULL
}
\keyword{internal}
