% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_plots.R
\name{brighten}
\alias{brighten}
\title{Brighten up a colour}
\usage{
brighten(col, factor)
}
\arguments{
\item{col}{Colour to be brightened up. Can be anything that is digestible by \code{\link{col2rgb}}.}

\item{factor}{Numeric The rate with which the color channel should be boosted. Must be in the range between 0 and 1.}
}
\value{
A character vector as returned by \code{\link{rgb}}.
}
\description{
We use this internal utility function to brighten up a specified colour (e.g., the swarm colour in \code{\link{apa_beeplot}}).
}
\keyword{internal}
