% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cite_r.R
\name{cite_r}
\alias{cite_r}
\title{Cite R and R-packages}
\usage{
cite_r(file = NULL, prefix = "R-", footnote = FALSE, pkgs = "all")
}
\arguments{
\item{file}{Character. The path and name of the \code{.bib}-file holding the references.}

\item{prefix}{Character. The prefix used for all R-package reference handles.}

\item{footnote}{Logical. Indicates if packages should be cited in a footnote. Ignored if no package information
is available.}

\item{pkgs}{Character. Vector of package names to cite. If \code{pkgs = "all"} (default) packages in \code{file}
are cited.}
}
\value{
If \code{footnote = FALSE} a character string is returned and a named list with the elements \code{r}
   and \code{pkgs} otherwise.
}
\description{
Creates a character string citing R and  R-packages.
}
\details{
If \code{footnote = FALSE} a character string citing R and R-packages including version
   numbers is returned. Otherwise a named list with the elements \code{r} and \code{pkgs} is returned. The
   former element holds a character string citing R and a reference to a footnote; the latter element contains
   a character string for the footnote citing R-packages. For correct rendering, the footnote string needs
   to be a separate paragraph.
}
\examples{
cite_r()
}
\seealso{
\code{\link{r_refs}}
}

