% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{default_label}
\alias{default_label}
\alias{default_label.data.frame}
\title{Set default variable labels from column names}
\usage{
default_label(x, ...)

\method{default_label}{data.frame}(x)
}
\arguments{
\item{x}{A \code{data.frame}}
}
\value{
Returns a \code{data.frame} with labelled columns. Labels are preserved (if already specified), otherwise generated from column names.
}
\description{
We use this function internally to provide default variable for all columns in a data.frame from column names.
}
\keyword{internal}
