% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\docType{methods}
\name{default_label}
\alias{default_label}
\alias{default_label,data.frame-method}
\title{Set default variable labels from column names}
\usage{
default_label(object)

\S4method{default_label}{data.frame}(object)
}
\arguments{
\item{x}{A \code{data.frame}}
}
\value{
Returns a \code{data.frame} with labelled columns. Labels are preserved (if already specified), otherwise generated from column names.
}
\description{
We use this function internally to provide default variable for all columns in a data.frame from column names.
}
\details{
#' @keywords internal

setMethod(
  "assign_annotation"
  , signature = c(object = "vector", value = "character", name = "character")
  , definition = function(object, value, name){
    object <- new(
      paste0("annotated_", class(object))
      , annotation <- new("vector_annotation")
    )
    object@label <- object@annotation@label
    object# return
  }
)
}
\keyword{internal}
