% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape symbols for LaTex output}
\usage{
escape_latex(x, newlines = FALSE, spaces = FALSE)
}
\arguments{
\item{x}{Character.}

\item{newlines}{Logical. Determines if \code{\\n} are escaped.}

\item{spaces}{Logical. Determines if multiple spaces are escaped.}
}
\description{
This function is a copy of the non-exported function \code{escape_latex} from the \pkg{knitr} package.
\emph{This function is not exported.}
}
\examples{
\dontrun{
in_paren <- TRUE # Taken from printnum()
validate(in_paren, check_class = "logical", check_length = 1)
validate(in_paren, check_class = "numeric", check_length = 1)
}
}

