% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{fast_aggregate}
\alias{fast_aggregate}
\title{Aggregate data much faster using dplyr}
\usage{
fast_aggregate(data, factors, dv, fun, na.rm = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data.}

\item{factors}{Character. Data is intended to be aggregated by these factors.}

\item{dv}{Character. The dependent variable that is aggregated. All variables in \code{data} that contain this string will be aggregated separately.}

\item{fun}{The function used for aggregation.}

\item{na.rm}{Logical. Is passed to \code{fun} and specifies if missing values will be removed.}
}
\description{
This is a convenience wrapper for aggregating your data using dplyr functions that tend to be much faster than the usual \code{aggregate} command.
It is also easy to call from within a function.
}
\examples{
NULL
}

