% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{fast_aggregate}
\alias{fast_aggregate}
\title{Aggregate data much faster using dplyr}
\usage{
fast_aggregate(data, factors, dv, fun)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data.}

\item{factors}{Character. A vector of factor names to aggregate data by.}

\item{dv}{Character. The dependent variable to aggregate. All variables in \code{data} that contain this substring
will be aggregated separately.}

\item{fun}{Closure. The function used for aggregation.}
}
\description{
This is a convenience wrapper for aggregating your data using dplyr functions that tend to be much faster than the
usual \code{aggregate} command. It is also easy to call from within a function. \emph{This function is not exported.}
}
\examples{
NULL
}

