% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_paren.R
\name{in_paren}
\alias{in_paren}
\title{Replace parentheses with brackets}
\usage{
in_paren(x)
}
\arguments{
\item{x}{Character. Single character or list of characters.}
}
\value{
An object of the same type as \code{x}.
}
\description{
Takes a single character or a list of characters and replaces paretheses with brackets. Can be used to
prepare a string of statistics (e.g. containing degrees of freedom) for reporting within paretheses.
}
\examples{
t_stat <- t.test(extra ~ group, data = sleep)
t_test_res <- apa_print(t_stat)
in_paren(t_test_res$stat)
}
\seealso{
apa_print
}

