% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\name{localize}
\alias{localize}
\title{Lookup Table for Genearted Words and Phrases}
\usage{
localize(x)
}
\arguments{
\item{x}{Integer. Locale.}
}
\description{
Some words and phrases used throughout a papaja manuscript are automatically
generated and need to vary when the locale of a document is changed. This
function returns the words and phrases by language.
}
\keyword{internal}
