% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tables.R
\docType{data}
\name{lookup_names}
\alias{lookup_names}
\alias{lookup_labels}
\title{Lookup Tables for Column Names and Variable Labels}
\format{
An object of class \code{character} of length 73.

An object of class \code{character} of length 68.
}
\usage{
lookup_names

lookup_labels
}
\description{
\code{apa_print()} converts many statistical output objects to output objects that
contain an \code{apa_results_table/data.frame} with standardized column names and
variable labels. For this purpose, it relies on the lookup tables provided
here.
}
\keyword{internal}
