% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_table.R
\name{merge_tables}
\alias{merge_tables}
\title{Merge tables in list}
\usage{
merge_tables(x, empty_cells, row_names, added_stub_head)
}
\arguments{
\item{x}{List. A named list containing one or more \code{matrix} or \code{data.frame}.}

\item{empty_cells}{Character. String to place in empty cells; should be \code{""} if the target document is LaTeX and
\code{"&nbsp;"} if the target document is Word.}

\item{row_names}{Logical. Vector of boolean values specifying whether to print column names for the corresponding list
element.}

\item{added_stub_head}{Character. Vector of names for first unnamed columns. See \code{\link{apa_table}}.}
}
\description{
Takes a list of containing one or more \code{matrix} or \code{data.frame} and merges them into a single table.
\emph{This function is not exported and currently unused.}
}
\examples{
NULL
}
\seealso{
\code{\link{apa_table}}
}

