% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_plots.R
\name{points.matrix}
\alias{points.matrix}
\title{Matrix method for points}
\usage{
\method{points}{matrix}(x, y, type = "p", ...)
}
\description{
Internal function for convenient plotting of multiple points.
}
\keyword{internal}
