% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify_terms}
\alias{prettify_terms}
\title{Prettify term names}
\usage{
prettify_terms(x, standardized = FALSE)
}
\arguments{
\item{x}{Character. Vector of term-names to be prettified}
}
\description{
Remove parentheses, replace colons with \code{$\\times$}. Useful to prettify term names in \code{apa_print()} tables. \emph{This function is not exported.}
}
\examples{
prettify_terms(c("material:valence", "material:valence:measurement"))
}

