% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interval.R
\name{print_interval}
\alias{print_interval}
\alias{print_interval.data.frame}
\alias{print_interval.numeric}
\alias{print_interval.matrix}
\alias{print_confint}
\alias{print_hdint}
\title{Create interval strings}
\usage{
print_interval(
  x,
  ...,
  conf_level = NULL,
  interval_type = NULL,
  use_math = FALSE
)

\method{print_interval}{data.frame}(x, ...)

\method{print_interval}{numeric}(
  x,
  y = NULL,
  conf_level = NULL,
  interval_type = NULL,
  use_math = FALSE,
  ...
)

\method{print_interval}{matrix}(
  x,
  conf_level = NULL,
  interval_type = NULL,
  use_math = FALSE,
  ...
)

print_confint(x, ..., interval_type = "CI")

print_hdint(x, ..., interval_type = "HDI")
}
\arguments{
\item{x}{Numeric. A vector (of length 2, unless `y` is also specified) with,
a two-column `matrix`, or a `data.frame`, which can coerced to a
`matrix`.}

\item{...}{
  Arguments passed on to \code{\link[=printnum]{printnum}}
  \describe{
    \item{\code{}}{}
  }}

\item{conf_level}{Numeric. Confidence level of the interval. Ignored if level can be infered from attributes of `x`, see Details.}

\item{interval_type}{Character. Abbreviation indicating the type of interval
estimate, e.g. `CI`.}

\item{use_math}{Logical. Indicates whether to insert `$` into the
output so that `Inf` or scientific will be rendered correctly.}

\item{y}{Numeric. A vector of the same length as `x`.}
}
\value{
A singel interval is returned as a `character` vector of length 1;
   multiple intervals are returned as a named `list` of `character` vectors
   of length 1.
}
\description{
Creates a character string to report an interval.
\emph{This function is not exported.}
}
\details{
If possible the confidence level of the interval is inferred from
   attributes of `x`. For a vector of length 2, the attribute `conf.level` is
   is consulted; for a `matrix` or `data.frame` the column names are used,
   if they are of the format "2.5 \%" and "97.5 \%".

   If `x` is a `matrix` or `data.frame` the row names are used as names for
   the returned `list` of intervals.
}
\examples{
\dontrun{
print_confint(1, 2, conf_level = 0.95)
print_confint(c(1, 2), conf_level = 0.95)
print_confint(matrix(c(1, 2), ncol = 2), conf_level = 0.95)
print_confint(confint(lm(cars)))
print_confint(confint(lm(cars)), digits = 3)
}
}
\seealso{
\code{\link{printnum}}
}
\keyword{internal}
