% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_interval}
\alias{print_interval}
\title{Create interval estimate string}
\usage{
print_interval(x, conf_level = NULL, use_math = FALSE, interval_type,
  ...)
}
\arguments{
\item{x}{Numeric. Either a \code{vector} of length 2 with attribute \code{conf.level} or a two-column \code{matrix}
and confidence region bounds as column names (e.g. "2.5 \%" and "97.5 \%") and coefficient names as row names.}

\item{conf_level}{Numeric. Vector of length 2 giving the lower and upper bounds of the confidence region in case
they cannot be determined from column names or attributes of \code{x}.}

\item{use_math}{Logical. Indicates whether to insert \code{$} into the output so that \code{Inf} or scientific
notation is rendered correctly.}

\item{...}{Arguments passed on to \code{printnum}
\describe{
  \item{x}{Numeric. Can be either a single value, vector, or matrix.}
}}
}
\description{
Creates a character string to report an interval estimate. \emph{This function is not exported.}
}
\examples{
\dontrun{
print_confint(c(1, 2), conf_level = 0.95)
}
}
\seealso{
\code{\link{printnum}}
}
\keyword{internal}
