% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_comp.R
\name{print_model_comp}
\alias{print_model_comp}
\title{Format statistics model comparisons (APA 6th edition)}
\usage{
print_model_comp(x, models = NULL, ci = NULL, boot_samples = 1000,
  in_paren = FALSE)
}
\arguments{
\item{x}{Data.frame. A \code{data.frame} of class \code{apa_variance_table} as returned by \code{\link{arrange_anova}}.}

\item{models}{List. List containing fitted \code{lm}- objects that were compared using \code{anova()}. If the list is named, element names are used as model names in the ouptut object.}

\item{ci}{Numeric. Confidence level for the confidence interval for \eqn{\Delta R^2} if \code{x} is a model comparison object of class \code{anova}. If \code{ci = NULL} no confidence intervals are estimated.}

\item{boot_samples}{Numeric. Number of bootstrap samples to estimate confidence intervals for \eqn{\Delta R^2} if \code{x} is a model comparison object of class \code{anova}; ignored if \code{ci = NULL}.}

\item{in_paren}{Logical. Indicates if the formated string will be reported inside parentheses. See details.}
}
\value{
A named list containing the following components:

   \describe{
     \item{\code{stat}}{A named list of character strings giving the test statistic, parameters, and \emph{p}
         value for each factor.}
     \item{\code{est}}{A named list of character strings giving the effect size estimates for each factor.} % , either in units of the analyzed scale or as standardized effect size.
     \item{\code{full}}{A named list of character strings comprised of \code{est} and \code{stat} for each factor.}
     \item{\code{table}}{A data.frame containing the complete ANOVA table, which can be passed to \code{\link{apa_table}}.}
   }
}
\description{
This function is the workhorse of the \code{apa_print.anova} for model comparisons. It takes a \code{data.frame}
of class \code{apa_model_comp} and produces strings to report the results in accordance with APA manuscript
guidelines. \emph{This function is not exported.}
}
\examples{
   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   anova_table <- arrange_anova(summary(npk_aov))
   print_anova(anova_table)
}
\seealso{
\code{\link{arrange_anova}}, \code{\link{apa_print.aov}}
}

