% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_num.R
\name{print_scientific}
\alias{print_scientific}
\title{Typeset scientific notation}
\usage{
print_scientific(x)
}
\arguments{
\item{x}{Character.}
}
\description{
Typesets scientific notation of numbers into properly typeset math strings.
}
\examples{
papaja:::print_scientific("1.25e+04")
}
\keyword{internal}
