% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printnum.R
\name{printnum}
\alias{printnum}
\alias{printp}
\title{Prepare numeric values for printing}
\usage{
printnum(x, gt1 = TRUE, zero = TRUE, margin = 1, na_string = "", ...)

printp(x, na_string = "")
}
\arguments{
\item{x}{Numeric. Can be either a single value, vector, or matrix.}

\item{gt1}{Logical. Indicates if the absolute value of the statistic can, in principal, greater than 1.}

\item{zero}{Logical. Indicates if the statistic can, in principal, be 0.}

\item{margin}{Integer. If \code{x} is a matrix, the function is applied either across rows (\code{margin = 1})
or columns (\code{margin = 2}).}

\item{na_string}{Character. String to print if element of \code{x} is \code{NA}.}

\item{...}{Further arguments that may be passed to \code{\link{formatC}}}

\item{digits}{Integer. Number of decimal places to print.}
}
\description{
Converts numeric values to character strings for reporting.
}
\details{
If \code{x} is a vector, \code{digits}, \code{gt1}, and \code{zero} can be vectors
   according to which each element of the vector is formated. Parameters are recycled if length of \code{x}
   exceeds length of the parameter vectors. If \code{x} is a matrix, the vectors specify the formating
   of either rows or columns according to the value of \code{margin}.
}
\section{Functions}{
\itemize{
\item \code{printp}: Convenience wrapper for \code{printnum} to print p-values with three decimal places.
}}
\examples{
printnum(1/3)
printnum(1/3, gt1 = FALSE)
printnum(1/3, digits = 5)

printnum(0)
printnum(0, zero = FALSE)

printp(0.0001)
}

