% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printnum.R
\name{printp}
\alias{printp}
\title{Prepare numeric values for printing as p-value}
\usage{
printp(x, na_string = "")
}
\arguments{
\item{x}{Numeric. Can be either a single value, vector, or matrix.}

\item{na_string}{Character. String to print if element of \code{x} is \code{NA}.}
}
\description{
Convenience wrapper for \code{printnum.numeric} to print p-values with three decimal places.
}
\examples{
printnum(0.05)
printnum(0.0005)
printnum(0.99999999)
}
