% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_letter_format.R
\name{quote_from_tex}
\alias{quote_from_tex}
\title{Quote from TeX document}
\usage{
quote_from_tex(x, file)
}
\arguments{
\item{x}{Character. One or more quote labels.}

\item{file}{Character. Path to LaTeX file from which to quote.}
}
\value{
A character vector of LaTeX document text of class \code{knit_asis},
see \code{\link[knitr:asis_output]{knitr::asis_output()}}.
}
\description{
Includes a labelled quote from a LaTeX document 'asis'.
}
\details{
Searches the LaTeX document specified in \code{file} for labelled
quotes, i.e. paragraphs that are enclosed in \verb{\% <@~\{#quote-label\}} and
\verb{\% ~@>\}} tags in LaTeX comments on separate lines. The labelled quote is
then inserted and rendered \code{asis}.
}
