% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_letter_format.R
\name{quote_from_tex}
\alias{quote_from_tex}
\title{Quote from TeX document}
\usage{
quote_from_tex(x, file)
}
\arguments{
\item{x}{Character. One or more quote labels.}

\item{file}{Character. Path to LaTeX file from which to quote.}
}
\description{
Includes a labelled quote from a LaTeX document 'asis'.
}
\details{
Searches the LaTeX document for a labelled quotes preceeded and
  followed by \code{\% <@~{#quote-label}} and \code{\% ~@>} tags in LaTeX
  comments
}
