% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r_refs.R
\name{r_refs}
\alias{r_refs}
\title{Create a reference file for R and its packages}
\usage{
r_refs(file, prefix = "R-", ...)
}
\arguments{
\item{file}{Character. The path and name of the file to be created or updated.}

\item{prefix}{Character. A prefix to be used for all R-package reference handles.}

\item{...}{Further arguments to pass to \code{knitr::write_bib}.}
}
\description{
Creates a .bib-reference file for the used R version and all attached R-packages, so they can be cited
in a Markdown-document using pandoc or LaTeX.
}
\details{
This function is a wrapper for \code{write_bib} from the \pkg{knitr} package. If a file exists
   at the specified location, the function reads the file and appends missing citation information to the
   end of the file. It is recommended to create a new .bib-file dedicated to R-related references and
   adding it to the \code{bibliography} parameter in the document's yaml-header.
}
\examples{
r_refs("test_bibliography.bib") # Creates file in current working directory
}
\seealso{
\code{\link[knitr]{write_bib}}
}

