% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_refs.R
\name{r_refs}
\alias{create_bib}
\alias{r_refs}
\title{Create a reference file for R and its packages}
\usage{
r_refs(file, append = TRUE, prefix = "R-", type_pref = c("Article",
  "Book"), tweak = TRUE)

create_bib(x, file, append = TRUE, prefix = "R-", type_pref = c("Article",
  "Book"), tweak = TRUE)
}
\arguments{
\item{file}{Character. The path and name of the file to be created or updated.}

\item{append}{Logical. Indicates if existing \code{.bib}-files should be complemented or overwritten.
See details.}

\item{prefix}{Character. A prefix to be used for all R-package reference handles.}

\item{type_pref}{Character. A vector of BibTeX entry types in the order to look for them in packages
CITATION file. See details.}

\item{tweak}{Logical. Indicates whether to fix some known problems in the citations (based on
\code{\link[knitr]{write_bib}}).}

\item{x}{Character. Names of packages to include in bibliography.}
}
\description{
Creates a \code{.bib}-reference file for the installed R version and R-packages, so they can be cited
in a Markdown-document using pandoc or LaTeX.
}
\details{
\code{r_refs} is a wrapper for \code{\link{create_bib}} to create a bibliography for R and all attached or
   cached packages.

   By default, if a file exists at the specified location, \code{r_refs} reads the file and
   appends missing citation information to the end of the file (\code{\link{create_bib}} always overwrites
   existing files). This behavior may result in redundant package references in your \code{.bib}-file,
   which will be cited when using \code{\link{cite_r}}. It is recommended to use a \code{.bib}-file
   dedicated to R-references, setting \code{append = FALSE}, and adding the file to the
   \code{bibliography} parameter in the document's yaml-header.

   Beware that chunks loading packags should \emph{not} be cached. \code{r_refs} will make all packages
   loaded in cached chunks citable, but it won't know when you remove a package from a cached chunk. This
   can result in unused package references in your \code{.bib}-file that will be cited when using
   \code{\link{cite_r}}.

   If a package provides citation information in a CITATION file, a reference is selected based on the
   prefered order of reference types specified in \code{type_pref}. By default, articles are cited if
   available rather than books. If no reference of the specified types is available, the first reference
   is used. If multiple references of the prefered type are given in the CITATION file all are cited.
   Finally, if no CITATION file exists a reference is generated from the DESCRIPTION file by
   \code{\link{citation}}.
}
\examples{
NULL
}
\seealso{
\code{\link{cite_r}}, \code{\link[knitr]{write_bib}}, \code{\link{fetch_web_refs}}, \code{\link{citation}}, \code{\link{toBibtex}}
}

