% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_comments.R
\name{remove_comments}
\alias{remove_comments}
\title{Remove comments from R Markdown}
\usage{
remove_comments(x, file)
}
\arguments{
\item{x}{Character. Path to R Markdown file}

\item{file}{Character. Name of the new R Markdown file without comments.}
}
\description{
Removes markdown comments from an R Markdown file
}
