% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_appendix.R
\name{render_appendix}
\alias{render_appendix}
\title{Render appendix}
\usage{
render_appendix(x, bibliography = rmarkdown::metadata$bibliography,
  csl = rmarkdown::metadata$csl, quiet = TRUE, options = NULL)
}
\arguments{
\item{x}{Character. Input file name.}

\item{bibliography}{Character. Location of bibliography file(s) to use.}

\item{csl}{Character. Location of CSL file to use. Defaults to APA-style.}

\item{quiet}{Logical. Suppresses pandoc command line output; see \code{\link[rmarkdown]{render}}.}

\item{options}{Character. Vector of options passed to \code{\link[rmarkdown]{pandoc_convert}}.
If \code{FALSE} output will be included in the document.}
}
\description{
This functions renders an R Markdown document \emph{without} YAML header to a TeX fragment inside
an \code{appendix} environment.
}
\details{
By default \code{x} is converted to a TeX file which can be included in an R Markdown document
   as \code{include}:

   \preformatted{
   output:
     pdf_document:
       include:
       after_body: appendix.tex
   }

   If \code{render_appendix} is called form an R Markdown document with a target document type other
   than a PDF file a Markdown document is created and a warning returned (includes are only supported
   in PDF documents).

   Default chunk options and hooks are set to those used in the R Markdown document from
   which \code{render_appendix} is called; otherwise defaults of \code{\link[rmarkdown]{md_document}} are
   used. It is, therefore, recommended to include \code{render_appendix} in your parent document.
}
