% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_appendix.R
\name{render_appendix}
\alias{render_appendix}
\title{Render appendix}
\usage{
render_appendix(x)
}
\arguments{
\item{x}{Character. Input file name.}
}
\description{
This functions renders an R Markdown to TeX fragment without preamble so it can be included as appendix.
}
\details{
Because pandoc currently does not support \code{includes} \code{after_body} this function has no effect
when rendering word documents.
}

