% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_letter_format.R
\name{revision_letter_pdf}
\alias{revision_letter_pdf}
\title{Revision Letter}
\usage{
revision_letter_pdf(keep_tex = TRUE, ...)
}
\arguments{
\item{keep_tex}{Logical. Whether to keep the intermediate tex file used in
the conversion to PDF.}

\item{...}{
  Arguments passed on to \code{\link[bookdown:html_document2]{bookdown::pdf_document2}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Template for creating a journal revision letters.
}
\examples{
NULL
}
\seealso{
\code{\link[bookdown]{html_document2}}
}
