% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_letter_format.R
\name{revision_letter_pdf}
\alias{revision_letter_pdf}
\title{Journal revision letter}
\usage{
revision_letter_pdf(keep_tex = TRUE, ...)
}
\arguments{
\item{keep_tex}{Logical. Logical. Keep the intermediate tex file used in the conversion to PDF.}

\item{...}{Further arguments to pass to \code{\link[bookdown]{pdf_document2}}}
}
\description{
Template for creating an journal revision letters.
}
\examples{
NULL
}
\seealso{
\code{\link[bookdown]{pdf_document2}}
}
