% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_terms}
\alias{sanitize_terms}
\title{Sanitize term names}
\usage{
sanitize_terms(x, standardized = FALSE)
}
\arguments{
\item{x}{Character. Vector of term-names to be sanitized.}

\item{standardized}{Logical. If \code{TRUE} the name of the function \code{\link{scale}} will be
removed from term names.}
}
\description{
Remove characters from term names that will be difficult to adress using the \code{$}-operator. \emph{This function is
not exported.}
}
\examples{
\dontrun{
sanitize_terms(c("(Intercept)", "Factor A", "Factor B", "Factor A:Factor B", "scale(FactorA)"))
}
}

