% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa_barplot.R
\name{se}
\alias{se}
\title{Standard errors}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{\code{numeric}. A vector of observations from your dependent variable.}

\item{na.rm}{\code{logical}. Shall missing values (i.e., \code{NA}s) be removed?}
}
\description{
Returns the standard error of a vector of observations
}

