% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{se}
\alias{se}
\title{Standard errors}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric. A vector of observations.}

\item{na.rm}{Logical. Specifies if missing values should be removed.}
}
\description{
Calculates the standard error of the mean
}

