% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apa.R
\name{theme_apa}
\alias{theme_apa}
\title{APA compatible ggplot2 theme}
\usage{
theme_apa(base_size = 12, base_family = "", box = FALSE)
}
\arguments{
\item{base_size}{Numeric. Base font size; other font sizes and margins are adjusted relative
to this.}

\item{base_family}{Character. Base font family.}

\item{box}{Logical. Indicates whether to draw a box around the plot.}
}
\description{
A ggplot2 theme with a white panel background, no grid lines, large axis and legend titles,
and increased text padding for better readability.
}
\details{
This theme is an adaptation of \code{\link[ggplot2]{theme_bw}}. In ggplot2, themes set the
   general aspect of the plot such as the color of the background, gridlines, the size and colour
   of fonts.
}
\examples{
 \dontrun{
   # Copied from ?ggtheme
   p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour = factor(gear))) + facet_wrap(~ am)
   p
   p + theme_apa()
 }
}
