% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printnum.R
\name{typeset_scientific}
\alias{typeset_scientific}
\title{Typset scientific notation}
\usage{
typeset_scientific(x)
}
\arguments{
\item{x}{Character.}
}
\description{
Typsets scientific notation of numbers into properly typeset math strings.
}
\examples{
papaja:::typeset_scientific("1.25e+04")
}
