% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\docType{methods}
\name{variable_label}
\alias{variable_label}
\alias{variable_label<-}
\alias{variable_label,annotated_vector-method}
\alias{variable_label,data.frame-method}
\alias{variable_label<-,annotated_vector,character-method}
\alias{variable_label<-,vector,character-method}
\alias{variable_label<-,factor,character-method}
\alias{variable_label<-,annotated_factor,character-method}
\alias{variable_label<-,data.frame,character-method}
\title{Assign or Extract Variable Labels}
\usage{
variable_label(object)

variable_label(object) <- value

\S4method{variable_label}{annotated_vector}(object)

\S4method{variable_label}{data.frame}(object)

\S4method{variable_label}{annotated_vector,character}(object) <- value

\S4method{variable_label}{vector,character}(object) <- value

\S4method{variable_label}{factor,character}(object) <- value

\S4method{variable_label}{annotated_factor,character}(object) <- value

\S4method{variable_label}{data.frame,character}(object) <- value
}
\arguments{
\item{object}{Either a vector or a \code{data.frame}.}

\item{value}{Character. The variable label(s) to be assigned. If \code{variable_label} is applied to a single vector,
this should be a length-one argument. If applied to a \code{data.frame}, \code{value} is required to be a \emph{named} vector.
Check the examples for details.}
}
\description{
Assign or extract variable labels of a \code{vector} \emph{or}
the columns (i.e., vectors) of a \code{data.frame}.
}
\keyword{internal}
