% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{variable_label}
\alias{variable_label}
\alias{variable_label.default}
\alias{variable_label.data.frame}
\alias{variable_label<-}
\alias{variable_label<-.default}
\alias{variable_label<-.data.frame}
\title{Assign or Extract Variable Labels}
\usage{
variable_label(object, ...)

\method{variable_label}{default}(object, ...)

\method{variable_label}{data.frame}(object, ...)

variable_label(object, ...) <- value

\method{variable_label}{default}(object, ...) <- value

\method{variable_label}{data.frame}(object, ...) <- value
}
\arguments{
\item{object}{Either a vector or a \code{data.frame}.}

\item{value}{Character. The variable label(s) to be assigned. If \code{variable_label} is applied to a single vector,
this should be a length-one argument. If applied to a \code{data.frame}, \code{value} is required to be a \emph{named} vector.
Check the examples for details.}
}
\description{
Assign or extract variable labels of a \code{vector} \emph{or}
the columns (i.e., vectors) of a \code{data.frame}.
}
