% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{variable_label}
\alias{variable_label}
\alias{variable_label.default}
\alias{variable_label.data.frame}
\alias{variable_label<-}
\alias{variable_label<-.default}
\alias{variable_label<-.data.frame}
\title{Variable labels}
\usage{
variable_label(x)

\method{variable_label}{default}(x)

\method{variable_label}{data.frame}(x)

variable_label(x) <- value

\method{variable_label}{default}(x) <- value

\method{variable_label}{data.frame}(x) <- value
}
\arguments{
\item{x}{Either a vector or a \code{data.frame} containing the variables you want to label.}

\item{value}{A vector with the variable label(s) to be assigned. Can be anything, e.g. character, expression, etc.}
}
\value{
If applied to a \code{data.frame}, \code{variable_label} returns a named vector containing all variable labels.
        If applied to a \code{vector}, a length-one \code{character}.
}
\description{
Functions used to assign and extract variable labels of a \code{vector} \emph{or}
the columns (i.e., vectors) of a \code{data.frame}.
}
\examples{
variable_label(npk) <- c("Block number", "Nitrogen", "Phosphate", "Potassium", "Yield")
variable_label(npk)
}
