% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_data.R
\docType{class}
\name{vector_annotation-class}
\alias{vector_annotation-class}
\title{Annotations for Vectors}
\description{
This S4 class is used to represent annotations (i.e., meta-data) of a vector;
it is used as the meta-data container of the
\code{\link{annotated_vector-class}}.
It currently contains slots for a variable label and the unit of measurement,
but is intended to be incrementally extended.
}
\section{Slots}{

\describe{
\item{\code{label}}{Length-one character. A label for the vector.}

\item{\code{unit}}{Length-one character. The unit of measurement.}
}}

\keyword{internal}
